% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{add_1_to_age}
\alias{add_1_to_age}
\title{Increment age in a cow table}
\usage{
add_1_to_age(cows)
}
\arguments{
\item{cows}{See \link{cow_table}.}
}
\value{
A \link{cow_table}.
}
\description{
Add 1 to \code{age} variable in a cow table.
}
