% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{add_1_to_month}
\alias{add_1_to_month}
\title{Increment age and months_in_area in a cow table}
\usage{
add_1_to_month(cows)
}
\arguments{
\item{cows}{See \link{cow_table}.}
}
\value{
A \link{cow_table}.
}
\description{
Add 1 to \code{age} and \code{months_in_area} variable in a cow table.
}
