% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{add_newborns}
\alias{add_newborns}
\title{Add newborns to a cow_table}
\usage{
add_newborns(cows, area_table, i, max_cow_id, newborn_table, param_sim)
}
\arguments{
\item{cows}{See \link{cow_table}.}

\item{area_table}{See \link{area_table}.}

\item{i}{The number of months from the start of the simulation.}

\item{max_cow_id}{The ID of a cow at the last non-empty row of \code{cows}.}

\item{newborn_table}{A result of \code{\link[=setup_newborn_table]{setup_newborn_table()}}.}

\item{param_sim}{A list which combined \link{param}, a result of \code{\link[=process_param]{process_param()}} and a result of \code{\link[=calc_param]{calc_param()}}.}
}
\value{
A list consisted of two elements: \code{cows} and \code{max_cow_id}.
}
\description{
Add newborns to a cow_table
}
