% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{area_list}
\alias{area_list}
\alias{setup_tie_stall_table}
\title{Setup of \code{tie_stall_table}}
\usage{
setup_tie_stall_table(init_cows, areas)
}
\arguments{
\item{init_cows}{The element \code{init_cows} of a result of \code{\link[=setup_cows]{setup_cows()}}.}

\item{areas}{See \link{area_table}.}
}
\value{
A list composed of \link{tie_stall_table} and NA.
}
\description{
Make chamber matrix, which indicates in which chamber each cow is.
Cows kept in free-stall or paddock are not shown in this matrix.
}
\seealso{
\link{setup_rp_table} \link{tie_stall_table} \link{setup_cows} \link{setup_area_table}
}
