% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{areas}
\alias{areas}
\alias{setup_tie_stall_table}
\title{Setup of \code{tie_stall_table}}
\usage{
setup_tie_stall_table(area_table)
}
\arguments{
\item{area_table}{See \link{area_table}.}
}
\value{
A list composed of \link{tie_stall_table} and NA.
}
\description{
Make chamber matrix, which indicates in which chamber each cow is.
Cows kept in free-stall or paddock are not shown in this matrix.
}
\seealso{
\link{setup_rp_table} \link{tie_stall_table} \link{setup_cows} \link{setup_area_table}
}
