% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barn.R
\name{assign_cows}
\alias{assign_cows}
\title{Assign cows to area_list according to cow_table.}
\usage{
assign_cows(cows, area_list, area_assignment)
}
\arguments{
\item{cows}{See \link{cow_table}.}

\item{area_list}{See \link{setup_areas}.}

\item{area_assignment}{See \code{\link[=calculate_area_assignment]{calculate_area_assignment()}}.}
}
\value{
A \link{area_list}.
}
\description{
Assign \code{chamber_id} to cows allocated to tie-stall barns.
}
\note{
This function assign \code{cow_id} just for \code{area_list}. Assignment of \code{chamber_id} in \code{cows} must be done by \link{assign_chambers} before using this function.
}
