% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{assign_newborns}
\alias{assign_newborns}
\title{Assign \code{chamber_id} to newborns}
\usage{
assign_newborns(cows, area_table, area_list)
}
\arguments{
\item{cows}{See \link{cow_table}.}

\item{area_table}{See \link{area_table}.}

\item{area_list}{See \link{setup_areas} and \link{tie_stall_table}.}
}
\description{
Assign \code{chamber_id} to newborns
}
