% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\docType{data}
\name{barn_table}
\alias{barn_table}
\alias{a_chamber}
\title{Barn data.table to store status of barns}
\format{\link[data.table:data.table]{data.table::data.table}}
\usage{
a_chamber
}
\description{
\code{barn_table} is a \link[data.table:data.table]{data.table::data.table} to store status of barns.
Each barn have each \code{barn_table}.
The rows are consists of \code{a_chamber}, which indicates one chamber in a barn.
}
\details{
\code{chamber_id}, \code{is_edge}, and \code{group_id} are fixed. Values will not be changed while a simulation.
Other variables are flexible. Values will may be changed while a simulation.
\itemize{
\item \code{chamber_id}: Chamber ID.
\item \code{is_edge1}: Whether the chamber is at a right end of a lane.
\item \code{is_edge2}: Whether the chamer is at a left end of a lane.
\item \code{cow_id}: Cow ID in a lane.
\item \code{cow_status}: Infection status of the cow.
\item \code{is_exposed}:
TRUE = The cow is not isolated AND one or both of the neighbor cows is infected and not isolated.
FALSE = Both of the neibors are either uninfected or isolated.
NA = No cow in the chamber.
\item \code{is_isolated}: Whether the cow is isolated or not.
\item \code{hazard_ratio}: Hazard ratio of infection to the cow calculated based on neighbors' infection status.
\item \code{neighbor1_status}: Infection status of the neighbor in the right chamber.
\item \code{neighbor1_isolated}: Whether the cow in the right chamber is isolated or not.
\item \code{neighbor1_infectivity}:
TRUE when the neighbor in the right chamber is not isolated and is infectious. Otherwise, FALSE. NA is not allowed to this variable.
\item \code{neighbor2_status}, \code{neighbor2_isolated}, \code{neighbor2_infectivity}: Variables about the neighbor in the left chamber.
}
}
\seealso{
\link{cow_table} \link{rp_table}
}
\keyword{datasets}
