% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blvibmjp-package.R, R/blvibmjp.R
\docType{package}
\name{blvibmjp-package}
\alias{blvibmjp}
\alias{blvibmjp-package}
\title{blvibmjp: Simulation model of BLV infection in a Japanese dairy herd.}
\description{
This package provides functions to construct a individual based model of simulation of Bovine Leukemia virus in a tipical Japanese dairy herd.

This package provides functions to construct a individual based model of simulation of Bovine Leukemia virus in a tipical Japanese dairy herd.
}
\author{
\strong{Maintainer}: Yuri Fujimoto \email{fmsan51@gmail.com}


\strong{Maintainer}: Yuri Fujimoto \email{fmsan51@gmail.com}

}
