% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{calc_ai}
\alias{calc_ai}
\alias{calc_ai_list}
\alias{calc_first_ai_list}
\alias{calc_heat}
\title{Calculate the number of conducted AI}
\usage{
calc_ai_list(possible_heat, param_sim)

calc_first_ai_list(possible_heat, param_sim)

calc_heat(possible_heat, calculated_ai)
}
\arguments{
\item{possible_heat}{A list consisted of day of possible heats in a month.}

\item{param_sim}{A list which combined \link{param}, a result of \code{\link[=process_param]{process_param()}} and a result of \code{\link[=calc_param]{calc_param()}}.}

\item{calculated_ai}{A list consisted of \code{"succeeded_ai"} and \code{"detected_heat"}.}
}
\description{
Calculate the number of conducted AI
}
