% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barn.R
\name{calc_infection_in_barns}
\alias{calc_infection_in_barns}
\title{Calculate infection in barns}
\usage{
calc_infection_in_barns(cows, i, month, area_table, areas, param_sim)
}
\arguments{
\item{cows}{See \link{cow_table}.}

\item{i}{The number of months from the start of the simulation.}

\item{month}{The current month (1, 2, ..., 12).}

\item{area_table}{See \link{area_table}.}

\item{areas}{See \link{setup_areas} and \link{tie_stall_table}.}

\item{param_sim}{A list which combined \link{param}, a result of \code{\link[=process_param]{process_param()}} and a result of \code{\link[=calc_param]{calc_param()}}.}
}
\value{
A \link{cow_table}.
}
\description{
Calculate infection in barns depending on barn type (tied or freed)
}
