% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{calc_param}
\alias{calc_param}
\title{Calculate parameters necessary to the simulation.}
\usage{
calc_param(param, modification = NULL)
}
\arguments{
\item{param}{See \link{param}.}

\item{modification}{A list used to overwrite the defaut parameter like \code{list(parameter_name = new_value, ...)}.}
}
\value{
A parameter list.
}
\description{
Calculate parameters to the simulation and overwrite the default setting if necessary.
}
\details{
Parameters processed by \code{\link[=process_param]{process_param()}} are deteministic. Parameters calculated by \code{\link[=calc_param]{calc_param()}} are stochastic.
}
\seealso{
\code{\link[=calc_param_pre]{calc_param_pre()}} \code{\link[=calc_param_both]{calc_param_both()}}
}
