% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{calc_param_both}
\alias{calc_param_both}
\title{Calculate parameters used in both of prepare_data() and simulation}
\usage{
calc_param_both(param)
}
\arguments{
\item{param}{See \link{param}.}
}
\value{
A parameter list.
}
\description{
Calculate parameters which are in both of \code{\link[=prepare_data]{prepare_data()}} and simulation
}
\seealso{
\code{\link[=calc_param]{calc_param()}} \code{\link[=calc_param_pre]{calc_param_pre()}}
}
