% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{calc_param_pre}
\alias{calc_param_pre}
\title{Calculate parameters necessary to prepare_data()}
\usage{
calc_param_pre(param, modification = NULL)
}
\arguments{
\item{param}{See \link{param}.}

\item{modification}{A list used to overwrite the defaut parameter like \code{list(parameter_name = new_value, ...)}.}
}
\value{
A parameter list.
}
\description{
Calculate parameters which are used only in \code{\link[=prepare_data]{prepare_data()}} and overwrite the default setting if necessary.
}
\seealso{
\code{\link[=calc_param]{calc_param()}} \code{\link[=calc_param_both]{calc_param_both()}}
}
