% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{calc_params}
\alias{calc_params}
\alias{calc_params_once}
\alias{calc_params_repeatedly}
\title{Calculate parameters necessary to the simulation.}
\usage{
calc_params_once(modification = NULL)

calc_params_repeatedly(parameter, modification = NULL,
  i_simulation = NULL)
}
\arguments{
\item{modification}{A list like \code{list(parameter_name_to_overwrite = new_parameter)}.}

\item{parameter}{Parameter list.}

\item{i_simulation}{The iteration indicator of the simulation.}
}
\value{
An empty list (\code{calc_params_once()}) or an parameter list (\code{calc_params_repeatedly()})
}
\description{
Calculate parameters to the simulation and overwrite the default setting if necessary.
\code{calc_params_once()} calculate parameters at the start of the whole simulations. At that moment, it returns an empty list.
\code{calc_params_repeatedly()} calculate paramaters at the start of the each iteration of the simulation.
}
