% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_simulation_process.R
\name{calc_prev}
\alias{calc_prev}
\title{Calculate prevalence from \code{cow_table} or a csv file}
\usage{
calc_prev(cows = NULL, csv = NULL, type = c("prop", "count", "status"))
}
\arguments{
\item{cows}{See \code{cow_table}}

\item{csv}{Path to a csv file.}

\item{type}{\code{prop} means proportion of infected cows. \code{count} means the number of infected and non-infected cows. \code{status} means the number of \code{s} (non-infected), \code{ial} (asymptomatic), \code{ipl} (persistent lymphositosis) and \code{ebl} cows.}
}
\value{
A \link[data.table:data.table]{data.table} contains monthly prevalences.
}
\description{
Calculate monthly prevalences from \code{cow_table} or a csv file. Set either one of \code{cows} or \code{csv}.
}
