% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{change_area}
\alias{change_area}
\title{Check and move cows between areas}
\usage{
change_area(cows, area)
}
\arguments{
\item{cows}{See \link{cow_table}.}

\item{area}{See \link{area_table}.}
}
\value{
A \link{cow_table}.
}
\description{
Check and move cows between areas
}
