% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\docType{data}
\name{communal_pasture_table}
\alias{communal_pasture_table}
\alias{a_communal_pasture_use}
\title{A data.table to manage use of a communal pasture}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 1 rows and 4 columns.}
\usage{
a_communal_pasture_use
}
\description{
\code{communal_pasture_table} is a \link[data.table:data.table]{data.table} to manage use of a communal pasture by a farm.
To simulate a farm using a communal pasture, users must specify one \code{communal_pasture_table} consisted by following items before starting a simulation. To simulate a farm not using a communal pasture, users must not specify \code{communal_pasture_table}.
}
\details{
\itemize{
\item \code{area_out}, \code{area_back} (integer): Areas from where cows are send to a communal pasture and come back to the farm specified by \code{area_id} in \link{area_table}.
\item \code{condition_out}, \code{condition_back} (character): Condition that cows in the area are send to the communal pasture and come back to the farm. See \code{condition} part in \link{area_table} to know how to specify them.
}
}
\examples{
# Heifers are send to a communal pasture from a non-pregnant heifer barn (area_id = 2) at 13 months old and come back to a pregnant heifer barn (area_id = 3) eight month after they get pregnant.
# Delivered cows are send to the communal pasture from a delivered cow barn (area_id = 4) in April and come back to the same barn in October.
communal_pasture_use <- a_communal_pasture_use[rep(1, 2), ]
communal_pasture_use[,
 `:=`(area_out = c(2, 4),
      area_back = c(3, 4),
      condition_out = c("age == 20", "month == 4"),
      condition_back = c("months_from_pregnancy == 8", "month == 10"))]
}
\seealso{
\link{cow_table} \link{tie_stall_table} \link{area_table} \link{movement_table} \link{rp_table}
}
\keyword{datasets}
