% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\docType{data}
\name{cow_table}
\alias{cow_table}
\alias{a_new_calf}
\title{A data.table to store status of cows}
\format{\link[data.table:data.table]{data.table}}
\usage{
a_new_calf
}
\description{
\code{cow_table} is a \link[data.table:data.table]{data.table} to store status of cows.
Status of all the cows in a farm is stored in one \code{cow_table}.
The rows are consists of \code{a_new_calf}, which indicates one cow.
}
\details{
\code{date_xxx} variables are expressed as month. The month when the simulation starts is 0. eg. The \code{date_birth} of a cow which is 20 months old at the simulation start is -20.
\code{day_xxx} variables are days in the month.
\itemize{
\item \code{row_id}: Fixed.
\item \code{cow_id}
\item \code{age}: Age in month.
\item \code{stage}: One of "calf", "heifer", "milking" or "dry".
\item \code{sex}: One of "female" (Holstern), "male" (Holstein), "freemartin" (female Hostein), "f1-female" (Hostein x Japanese black), "f1-male" (Holstein x Japanese black), "black-female" (Japanese black), "black-male" (Japanese black).
\item \code{date_birth}: Date of birth.
\item \code{date_death}: Date of death. The value is NA while the cow is alive. It is set after the cow died.
\item \code{date_death_expected}: Expected date of death. It is calculated when a calf is born or when a cow is introduced to the herd.
\item \code{is_owned}: Whether the cow is owned by the simulated herd. It is set FALSE when the cow is sold or died.
\item \code{cause_removal}:
\itemize{
\item "died": culled or slaughtered.
\item "ebl": culled due to onset of the disease.
\item "sold": send to a market or another farm.
\item "will_die": (still alive and) will die.
\item "will_be_slaughtered": (still alive and) will be slaughtered.
}
\item \code{is_replacement}:
TRUE: The cow will be kept in the farm as replacement.
FALSE: The cow Will be sold to beef operations.
\item \code{is_introduced}
\item \code{is_in_common_ranch}: Whether the cow is in a common ranch at \code{i_month} (see below).
\item \code{is_grazed}
\item \code{parity}
\item \code{date_last_delivery}: 'Delivery' here includes abortions and stillbirths.
\item \code{date_got_pregnant}: NA means that the cow is open.
\item \code{date_dried}: NA means that the cow is milking.
\item \code{n_ai}: The number of AI from last delivery. The value is set as 0 when the cow got pregnant.
\item \code{day_heat}: Day in month of the NEXT heat.
\item \code{day_last_detected_heat}: Dey in month of the LAST heat.
\item \code{is_to_test_pregnancy}: Whether to be served for a pregnancy test (= AI was conducted to the cow and heats were not observed from then).
\item \code{infection_status}: One of "s" (susceptible = non-infected), "ial" (aleukemia), "ipl" (PL) and "ebl".
\item \code{is_detected}: Whether BLV infection is detected. \code{NA} for non-infected cows.
\item \code{date_ial}: The month when \code{infection_status} changes from "s" to "ial".
\item \code{date_ipl}: The month when \code{infection_status} changes from "ial" to "ipl".
\item \code{date_ipl_expected}: The expected month when \code{infection_status} changes from "ial" to "ipl". It is calculated when \code{infection_status} becomes "ial".
\item \code{date_ebl}: The month when \code{infection_status} changes from "ipl" to "ebl".
\item \code{date_ebl_expected}: The exppected month when \code{infection_status} changes from "ial" to "ebl". It is calculated when \code{infection_status} becomes "ial".
\item \code{cause_infection}: The reason why the cow is infected. It is set when the cow gets infected.
\itemize{
\item "initial": already be infected at the start of the start of a simulation.
\item "insects": by bloodsucking insects.
\item "contact": by direct contact with an infected cow.
\item "needles": by a contaminated needle.
\item "rp": by rectal palpation.
\item "vertical": vertical transimisson (intrauterine and by the dam milk).
\item "introduced": for a introduced cow which is infected before introduction (a home-bred cow infected from introduced cow is not categorized as this).
\item "comranch": infected at a communal ranch.
}
\item \code{susceptibility_ial_to_ipl}: Genetic susceptibility to disease progress (Ial -> Ipl).
\item \code{susceptibility_ipl_to_ebl}: Genetic susceptibility to disease progress (Ipl -> EBL).
\item \code{area_id}: Area ID.
\item \code{months_in_area}: The number of month a cow stayed in the current area.
\item \code{chamber_id}: ID of the chamber in which the cow kept for a cow in a tie-stall barn. \code{NA_real_} for a cow in a free-stall barn.
\item \code{is_isolated}: Whether the cow is isolated for a cow in a tie-stall barn. \code{NA_real_} for a cow in a free-stall barn.
\item \code{i_month}: The number of months past from the start of a simulation.
}
}
\seealso{
\link{tie_stall_table} \link{area_table} \link{movement_table} \link{rp_table}
}
\keyword{datasets}
