% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{cull_infected_cows}
\alias{cull_infected_cows}
\title{Cull infected cows}
\usage{
cull_infected_cows(cows, areas, i, area_table, param_sim)
}
\arguments{
\item{cows}{See \link{cow_table}.}

\item{areas}{See \link{tie_stall_table}.}

\item{i}{The number of months from the start of the simulation.}

\item{area_table}{See \link{area_table}.}

\item{param_sim}{A list which combined \link{param}, a result of \code{\link[=process_param]{process_param()}} and a result of \code{\link[=calc_param]{calc_param()}}.}
}
\value{
A \link{cow_table}.
}
\description{
Cull infected cows
}
