% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{days_per_month}
\alias{days_per_month}
\title{Days per month (365 / 12)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
days_per_month
}
\description{
Days per month (365 / 12)
}
\keyword{datasets}
