% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barn.R
\name{depreciate_is_ts}
\alias{depreciate_is_ts}
\title{Check whether a barn is tie-stall}
\usage{
depreciate_is_ts(param_area)
}
\arguments{
\item{param_area}{See \link{param_area}.}
}
\value{
A logical vector of length 4 which indicates whether barns are tie-stall or not.
}
\description{
Check whether a barn is tie-stall
}
