% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barn.R
\name{depreciate_make_ts_area}
\alias{depreciate_make_ts_area}
\title{Make a tie_stall_table from a cow_table}
\usage{
depreciate_make_ts_area(cows, n_x, n_y)
}
\arguments{
\item{cows}{See \link{cow_table}.}

\item{n_x}{The number of chambers in a lane.}

\item{n_y}{The number of lanes in a tie-stall barn.}
}
\value{
A \link{tie_stall_table}.
}
\description{
Make a \link{tie_stall_table} based on x, y coordinates in a \link{cow_table}.
}
