% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barn.R
\name{depreciated_find_empty_chamber}
\alias{depreciated_find_empty_chamber}
\title{Find empty chambers in a tie_stall_table}
\usage{
depreciated_find_empty_chamber(area, added_cows)
}
\arguments{
\item{area}{See \link{tie_stall_table}.}

\item{added_cows}{A \link{cow_table} consisted of cows to add to the area.}
}
\value{
A \link{tie_stall_table}.
}
\description{
Find empty chambers in a tie_stall_table and assign cows to chambers.
}
