% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{do_ai}
\alias{do_ai}
\title{Conduct AI and check chance of infection}
\usage{
do_ai(cows, i)
}
\arguments{
\item{cows}{See \link{cow_table}.}

\item{i}{The number of months from the start of the simulation.}
}
\value{
A \link{cow_table}
}
\description{
Conduct AI and check chance of infection
}
