% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{do_test}
\alias{do_test}
\title{Conduct BLV test}
\usage{
do_test(cows, month, param_sim)
}
\arguments{
\item{cows}{See \link{cow_table}.}

\item{month}{The month (1, 2, ..., 12).}

\item{param_sim}{A list which combined \link{param}, a result of \code{\link[=process_param]{process_param()}} and a result of \code{\link[=calc_param]{calc_param()}}.}
}
\value{
A \link{cow_table}.
}
\description{
Conduct BLV test
}
