% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{est_coef_inf_insects}
\alias{est_coef_inf_insects}
\title{Likelihood function of a coefficient for infection by insects}
\usage{
est_coef_inf_insects(coef, risks_inf_insects, prob_seroconv_insects)
}
\arguments{
\item{coef, risks_inf_insects, prob_seroconv_insects}{See the source code of \code{\link[=calc_param]{calc_param()}} for detail.}
}
\description{
A function used to estimate coefficient for risks_inf_insects in \code{\link[=calc_param]{calc_param()}}.
}
