% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{extract_owned_cows}
\alias{extract_owned_cows}
\title{Extract owned cows from a cow_table}
\usage{
extract_owned_cows(cows)
}
\arguments{
\item{cows}{See \link{cow_table}.}
}
\value{
A \link{cow_table}.
}
\description{
Extract owned cows from a cow_table
}
