% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\docType{data}
\name{farm_parameter}
\alias{farm_parameter}
\title{Parameters about farms which should be set by users}
\format{An object of class \code{list} of length 18.}
\usage{
farm_parameter
}
\description{
\itemize{
\item \code{csv}: Path to a csv file which contains cattle information.
\item \code{prop_female} (0-1): Proportion of female of newborns (NA: average in Hokkaido)
\item \code{prop_replacement} (0-1): Proportion of calvest to be replacements in female newborns (NA: average in Hokkaido)
\item \code{prop_died} (1-0): Proportion of dead cows in removed cows (Died / (Died + Slaughtered)) (NA: average in Hokkaido)
\item \code{prop_heat_detected} (0-1): Proportion of detected heats in total heats (NA: average in Hokkaido)
\item \code{calving_interval}: Calving interval in day (NA: average in Hokkaido)
\item \code{n_mean_ai}: Mean number of AI done until conception
\item \code{mean_age_first_ai}, \code{sd_age_first_ai}: Age (in month) of the first AI for heifers. $~Norm(mean, sd)$. (NA: mean = mode of Hokkaido, sd = \code{2 / qnorm(0.975)})
\item \code{mean_day_start_ai}, \code{sd_day_start_ai}: Day of first AI after a delivery. $~Norm(mean, sd)$. (NA: mean = mean of Hokkaido, sd = \code{10 / qnorm(0.975)})
\item \code{months_grazing} (1-12), \code{hours_grazing} (0-23): Months and hours for grazing (NA: no grazing)
\item \code{capacity_in_head} c(lower, upper): Lower/upper limit of the herd size. Set either this or \code{capacity_as_ratio} below.
\item \code{capacity_as_ratio} c(lower, upper): Lower/upper limit of the herd as ratio to the initial herd size (lower limit = \code{lower * initial_herd_size}, upper limit = \code{upper * initial_herd_size}). Set either this or \code{capacity_in_head} above. When both of \code{capacity_in_head} and \code{capacity_as_ratio} is NA, \code{capacity_as_ratio} is set to \code{c(0.9, 1.1)}.
\item \code{change_needles} (logical): whether use one needles for one cow (NA: TRUE)
\item \code{change_gloves} (logical): whether use one glove for one cow for rectal palpation (NA: TRUE)
\item \code{days_milking}: Length of milking period (in days) (NA: average in Hokkaido)
}
}
\keyword{datasets}
