% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_function.R
\name{heat_cycle}
\alias{heat_cycle}
\title{Heat cycle}
\usage{
heat_cycle(n_cows, stage, param_sim)
}
\arguments{
\item{n_cows}{The number of cows to which heat cycle should be calculated.}

\item{stage}{Stage (\code{stage} column of \link{cow_table} of cows.}

\item{param_sim}{A list which combined \link{param}, a result of \code{\link[=process_param]{process_param()}} and a result of \code{\link[=calc_param]{calc_param()}}.}
}
\value{
A numeric vector.
}
\description{
Heat cycle
}
