% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Test if a numeric vector is consisted of whole numbers or not}
\usage{
is.wholenumber(number)
}
\arguments{
\item{number}{numeric vector.}
}
\value{
A logical value.
}
\description{
Test if all the components of a numeric vector is whole number or not.
}
\examples{
blvibmjp:::is.wholenumber(3L)  # TRUE
blvibmjp:::is.wholenumber(3)  # TRUE
blvibmjp:::is.wholenumber(pi)  # FALSE

}
