% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is.wholenumbers}
\alias{is.wholenumbers}
\title{Test if a numeric vector is consisted of whole numbers or not}
\usage{
is.wholenumbers(number, na.rm = F)
}
\arguments{
\item{number}{numeric vector.}

\item{na.rm}{logical. Passed to \code{na.rm} of \code{\link[=all]{all()}}.}
}
\value{
A logical value.
}
\description{
Test if all the components of a numeric vector is whole number or not.
}
\examples{
blvibmjp:::is.wholenumbers(3L)  # TRUE
blvibmjp:::is.wholenumbers(3)  # TRUE
blvibmjp:::is.wholenumbers(pi)  # FALSE

}
