% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_function.R
\name{is_ai_started_heifer}
\alias{is_ai_started_heifer}
\title{Whether the first AIs for hifers are conducted}
\usage{
is_ai_started_heifer(ages, param_sim)
}
\arguments{
\item{ages}{Age of heifers.}

\item{param_sim}{A list which combined \link{param}, a result of \code{\link[=process_param]{process_param()}} and a result of \code{\link[=calc_param]{calc_param()}}.}
}
\value{
A logical vector.
}
\description{
Whether the first AIs for hifers are conducted
}
