% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{is_ai_started_milking}
\alias{is_ai_started_milking}
\title{Whether the first AIs for milking cows are conducted}
\usage{
is_ai_started_milking(n_month_from_delivery)
}
\arguments{
\item{n_month_from_delivery}{The month past from the last delivery.}
}
\value{
A logical vector
}
\description{
Whether the first AIs for milking cows are conducted
}
