% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{is_first_ai_successed}
\alias{is_first_ai_successed}
\alias{is_ai_successed}
\title{Whether a cow is concepted at the first or later AI}
\usage{
is_first_ai_successed(n_cows)

is_ai_successed(n_cows)
}
\arguments{
\item{n_cows}{The number of cows on which the first AI were conducted}
}
\value{
A logical vector
}
\description{
Whether a cow is concepted at the first or later AI
}
