% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_function.R
\name{is_dried}
\alias{is_dried}
\title{Whether a cow is dried}
\usage{
is_dried(months_from_delivery, param_calculated)
}
\arguments{
\item{months_from_delivery}{The number of months past from the last delivery.}

\item{param_calculated}{Return from \code{\link[=calc_param]{calc_param()}}.}
}
\value{
A logical vector.
}
\description{
Whether a cow is dried
}
