% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_function.R
\name{is_infected_by_colostrum}
\alias{is_infected_by_colostrum}
\title{Wheter newborns are infected by colostrum milk}
\usage{
is_infected_by_colostrum(status_mother, param_sim)
}
\arguments{
\item{status_mother}{The \code{infection_status} of dams.}

\item{param_sim}{A list which combined \link{param}, a result of \code{\link[=process_param]{process_param()}} and a result of \code{\link[=calc_param]{calc_param()}}.}
}
\value{
A logical vector.
}
\description{
Wheter newborns are infected by colostrum milk
}
