% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{is_infected_comranch}
\alias{is_infected_comranch}
\title{Whether cows are infected at a communal ranch}
\usage{
is_infected_comranch()
}
\value{
A logical vector
}
\description{
When a cow is come back from a communal ranch, the probability of infection at communal ranch was calculated.
}
