% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_function.R
\name{is_infected_in_exposed_chamber}
\alias{is_infected_in_exposed_chamber}
\title{Whether cows are infected in chambers next to infected cows in tie-stall barns}
\usage{
is_infected_in_exposed_chamber(n_cows, param_calculated)
}
\arguments{
\item{n_cows}{The number of cows.}

\item{param_calculated}{Return from \code{\link[=calc_param]{calc_param()}}.}
}
\value{
A logical vector.
}
\description{
Whether cows are infected in chambers next to infected cows in tie-stall barns
}
