% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_function.R
\name{is_infected_in_free_stall}
\alias{is_infected_in_free_stall}
\title{Whether cows are infected in free pastures}
\usage{
is_infected_in_free_stall(n_noinf, n_inf, month, param_sim)
}
\arguments{
\item{n_noinf}{The number of non-infected cows in a barn.}

\item{n_inf}{The number of infected cows in the barn.}

\item{month}{The current month (1, 2, ..., 12).}

\item{param_sim}{A list which combined \link{param}, a result of \code{\link[=process_param]{process_param()}} and a result of \code{\link[=calc_param]{calc_param()}}.}
}
\value{
A logical vector.
}
\description{
Whether cows are infected in free pastures
}
