% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{is_infected_introduced}
\alias{is_infected_introduced}
\title{Whether introduced cows are infected}
\usage{
is_infected_introduced()
}
\value{
A logical vector
}
\description{
Whether introduced cows are infected
}
