% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{is_infected_needles}
\alias{is_infected_needles}
\title{Whether cows are infected by contaminated needles}
\usage{
is_infected_needles(id_cow_s)
}
\arguments{
\item{id_cow_s}{}
}
\value{
A logical vector
}
\description{
Whether cows are infected by contaminated needles
}
