% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_function.R
\name{is_infected_vertical}
\alias{is_infected_vertical}
\title{Wheter newborns are infected vertically}
\usage{
is_infected_vertical(status_mother, param_calculated)
}
\arguments{
\item{status_mother}{The \code{infection_status} of dams.}

\item{param_calculated}{Return from \code{\link[=calc_param]{calc_param()}}.}
}
\value{
A logical vector.
}
\description{
Wheter newborns are infected vertically
}
