% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barns.R
\name{is_md_separated_in_ts}
\alias{is_md_separated_in_ts}
\title{Check whether milking cows and dry cows are separated}
\usage{
is_md_separated_in_ts(params)
}
\arguments{
\item{params}{A list of parameters. See \code{\link[=calc_params]{calc_params()}}.}
}
\value{
A logical value
}
\description{
Check whether milking cows and dry cows are separated
}
