% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_function.R
\encoding{UTF-8}
\name{is_replacement}
\alias{is_replacement}
\alias{set_prob_rep}
\title{Whether a newborn will be a replacement}
\usage{
set_prob_rep(n_delivered, param)

is_replacement(n_calves, herd_size, param_processed)
}
\arguments{
\item{n_delivered}{The number of delivered cows in a herd.}

\item{param}{See \link{param}.}

\item{n_calves}{The number of newborns.}

\item{herd_size}{The current herd size.}

\item{param_processed}{Return from \code{\link[=process_param]{process_param()}}.}
}
\value{
A logical vector or a numeric value.
}
\description{
Whether a newborn will be a replacement
}
\references{
「平成28年度 乳用種初生牛の経営に関する調査報告書」の「表40 調査対象経営の乳用種雌子牛の仕向け状況（規模別）」
}
