% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_function.R
\name{is_stillbirth}
\alias{is_stillbirth}
\title{Whether a delivery end up in stillbirth or abortion}
\usage{
is_stillbirth(parity, param_sim)
}
\arguments{
\item{parity}{Parity of dams.}

\item{param_sim}{A list which combined \link{param}, a result of \code{\link[=process_param]{process_param()}} and a result of \code{\link[=calc_param]{calc_param()}}.}
}
\value{
A logical vector.
}
\description{
Here, stillbirth means stillbirth and abortion.
}
