% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barn.R
\name{is_ts}
\alias{is_ts}
\title{Check wheter a barn is tie-stall}
\usage{
is_ts(param_group)
}
\arguments{
\item{param_group}{See \link{param_group}.}
}
\value{
A logical vector of length 4 which indicates whether barns are tie-stall or not.
}
\description{
Check wheter a barn is tie-stall
}
