% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_function.R
\name{age_die}
\alias{age_die}
\alias{longevity}
\alias{age_slaughtered}
\title{Expected age of death or slaugher}
\usage{
age_die(n_cows, param_sim)

age_slaughtered(n_cows, param_sim)

longevity(n_cows, param_sim)
}
\arguments{
\item{n_cows}{The number of cows to calculate expected age of death.}

\item{param_sim}{A list which combined \link{param}, a result of \code{\link[=process_param]{process_param()}} and a result of \code{\link[=calc_param]{calc_param()}}.}
}
\value{
A numeric vector.
}
\description{
Expected age of death or slaugher
}
