% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{months_progression}
\alias{months_progression}
\title{Calculate months to progress the disease}
\usage{
months_progression(susceptibility_ial_to_ipl, susceptibility_ipl_to_ebl, i)
}
\arguments{
\item{susceptibility_ial_to_ipl}{A numeric.}

\item{susceptibility_ipl_to_ebl}{A numeric.}

\item{i}{The number of months from the start of a simulation.}
}
\value{
A list consisted of a month
}
\description{
When a cow is infected, the number of months necessary to progress the disease is calculated at the month a cow is infected.
}
\details{
Probabilities to:  
- be newly infected (s->ial),  
- develop persistent lymphocytosis (ial->ipl), and  
- show symptom (ipl->ebl),
in the month.

Total infection risk = Infection risk in free-stall or grazing  
  * (hours_grazing / 24)  
  + Infection risk from neighborhood cows * (1 - hours_grazing / 24)
}
