% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{months_until_ebl_die}
\alias{months_until_ebl_die}
\title{The number of months until EBL cows die}
\usage{
months_until_ebl_die(rows_cow_overlooked)
}
\arguments{
\item{rows_cow_overlooked}{The row IDs of EBL-yet-non-detected cows.}
}
\value{
A numeric vector
}
\description{
The number of months until EBL cows die
}
