% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\docType{data}
\name{movement_table}
\alias{movement_table}
\alias{a_movement}
\title{A data.table to manage cows' movement between areas}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1 rows and 4 columns.
}
\usage{
a_movement
}
\description{
\code{movement_table} is a \link[data.table:data.table]{data.table} to manage cow movement among areas.
Users must specify one \code{movement_table} consisted of following items before starting a simulation.
}
\details{
\itemize{
\item \code{current_area} (integer): The current area a cow is kept specified by \code{area_id} in \link{area_table}.
\item \code{condition} (character): Condition that cows in the area move to the next area(s). Describe conditions by character which can be evaluated as logical (see Example). You can use following terms to specify \code{condition}:
\itemize{
\item \code{age}: Age in month. Use like \code{age == 20}.
\item \code{parity}: Parity. Use like \code{parity > 1}.
\item \code{months_from_delivery}, \code{months_from_pregnancy}, \code{months_from_dry}: The number of months from delivery, pregnancy or dry. Use like \code{months_from_delivery == 0} (this means the month a cow delivered).
\item \code{delivery}, \code{pregnancy}, \code{dry}: A shorthand form of \code{months_from_delivery == 0} and so on.
\item \code{dim}: Days in milking. Use like \code{dim > 100}.
\item \code{month}: Month in a year (1 = Jan, 2 = Feb, ...). Use like \code{month == 3}.
\item \code{stay}: The number of months for which a cow stayed in a area. Use like \code{stay == 3}.
}
\item \code{next_area} (list consisted of integer): The next area a cow will move to specified by \code{area_id} in \link{area_table}. You can specify multiple areas like \code{c(1:2, 4)}.
\item \code{priority} (list consisted of integer and/or numeric): The priority for \code{next_area}. Specify integer or numeric vector (for numeric vector, they must be summed to 1,) whose length is equal to \code{next_area}. If \code{priority} is set by integer, cows move to the area with highest \code{priority} (= nearest to 1) which is not full. If multiple areas have the same \code{priority}, cows are romdomly allocated to the areas. If \code{priority} is set by numeric which is summed to 1, \code{priority} is regarded as probability in accordance to which cows move to \code{next_area}.
}

If a cow meets multiple conditions, the condition in the fastest row will be used.
}
\examples{
movements <- a_movement[rep(1, 3), ]
movements[, `:=`(current_area = c(1L, 2L, 3L),
                 condition = c("age > 2", "delivery", "delivery"),
                 next_area = list(2L, 3:4, 3:4),
                 priority = list(NA, NA, c(1, 2)))]

}
\seealso{
\link{cow_table} \link{tie_stall_table} \link{area_table} \link{rp_table}
}
\keyword{datasets}
