% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_function.R
\name{n_month_to_progress}
\alias{n_month_to_progress}
\title{Calculate months to progress the disease}
\usage{
n_month_to_progress(
  susceptibility_ial_to_ipl,
  susceptibility_ipl_to_ebl,
  i,
  param_calculated
)
}
\arguments{
\item{susceptibility_ial_to_ipl}{A numeric.}

\item{susceptibility_ipl_to_ebl}{A numeric.}

\item{i}{The number of months from the start of a simulation.}

\item{param_calculated}{Return from \code{\link[=calc_param]{calc_param()}}.}
}
\value{
A list consisted of a month.
}
\description{
When a cow is infected, the number of months necessary to progress the disease is calculated at the month a cow is infected.
}
\details{
Probabilities to:
\itemize{
\item be newly infected (s->ial),
\item develop persistent lymphocytosis (ial->ipl), and
\item show symptom (ipl->ebl),
in the month.
}

Total infection risk = Infection risk in free-stall or grazing
\itemize{
\item (hours_grazing / 24)
}
\itemize{
\item Infection risk from neighborhood cows * (1 - hours_grazing / 24)
}
}
