% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_function.R
\name{n_newborn_per_dam}
\alias{n_newborn_per_dam}
\title{The number of newborns per dam}
\usage{
n_newborn_per_dam(n_dams, param_calculated)
}
\arguments{
\item{n_dams}{The number of dams.}

\item{param_calculated}{Return from \code{\link[=calc_param]{calc_param()}}.}
}
\value{
A numeric vector consisted of 1 and 2.
}
\description{
The number of newborns per dam
}
