% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\docType{data}
\name{param_area}
\alias{param_area}
\title{Parameters about areas which should be set by users}
\format{An object of class \code{list} of length 1.}
\usage{
param_area
}
\description{
\itemize{
\item \code{calf_area_priority}: Specify priority for calf areas. \code{NA} is allowed. For detail of \code{priority}, see \link{area_table}.
}
}
\seealso{
\link{param_simulation} \link{param_farm}
}
\keyword{datasets}
