% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\docType{data}
\name{param_area}
\alias{param_area}
\title{Parameters about barns which should be set by users}
\format{An object of class \code{list} of length 5.}
\usage{
param_area
}
\description{
\itemize{
\item \code{calf_area_id}: Numeric vector. Set the \code{area_id}(s) for newborn calves specified in \code{[area_table]}.
\item \code{calf_area_priority}: Specify priority for calf areas. \code{NA} is allowed. See explanation of \code{priority} in \code{[area_table]} for meaning of \code{priority}.
\item \code{xy_chamber}: The number of chambers per lane and the number of lanes.
\item \code{is_calf_separated}: Whether each calf is separated.
\item \code{is_milking_dry_separated}: (For a farm which miling and dry cows are kept in same tie-stall barn,) whether dry cows are separated from milking cows.
}
}
\seealso{
\link{param_simulation} \link{param_farm}
}
\keyword{datasets}
