% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\docType{data}
\name{param_simulation}
\alias{param_simulation}
\title{Parameters about a simulation which should be set by users}
\format{An object of class \code{list} of length 6.}
\usage{
param_simulation
}
\description{
\itemize{
\item \code{simulation_length}: Length of simulation (months). (default: 60)
\item \code{n_simulation}: The number of simulation. (default: 1)
\item \code{simulation_start}: The month simulation starts (1 = Jan, 2 = Feb, ...). (default: 1)
\item \code{input_csv}: Path to a csv file which contains cattle information.
\item \code{output_dir}: Directory to output files. (default: data/output)
\item \code{output_filename}: The name of the output files. (default: "simulation")
}
}
\seealso{
\link{param_farm} \link{param_group}
}
\keyword{datasets}
