% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\docType{data}
\name{PARAMS_GROUPS}
\alias{PARAMS_GROUPS}
\title{Parameters about barns which should be set by users}
\format{An object of class \code{list} of length 4.}
\usage{
PARAMS_GROUPS
}
\description{
\itemize{
\item \code{n_groups}: The number of barns
\item \code{xy_chambers}
}
}
\keyword{datasets}
