% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_simulation_process.R
\name{plot_infection_route}
\alias{plot_infection_route}
\title{Plot monthly infection routes nicely}
\usage{
plot_infection_route(path_to_csv, language = NULL, route_levels = NULL,
  route_labels = NULL, max_ylim = 100, title = NULL,
  legend_title = "Infection route", xlab = "Months in simulation",
  ylab = "Number of cattle", gray = F, area_color = NULL,
  border = F, border_color = NULL, font = NULL)
}
\arguments{
\item{path_to_csv}{Path to an output csv file.}

\item{language}{When set, plot title and labels are translated in this language. At present, only Japanese is implemented.}

\item{route_levels, route_labels}{See \link{redefine_route_levels}}

\item{max_ylim}{Upper limit of the y-axis of the plot.}

\item{title, legend_title, xlab, ylab}{Plot title, legend title, label for x-axis and label for y-axis.}

\item{gray}{When \code{TRUE}, a plot will be a grayscale image.}

\item{area_color}{Specify a color palette of a plot.}

\item{border}{When \code{TRUE}, each area in a plot will be surrounded by border.}

\item{border_color}{Specify a color palette for the border.}

\item{font}{Font in a plot. The default is "Meiryo" for Windows and "Hiragino Kaku Gothic Pro" for the other OS.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} plot.
}
\description{
Plot monthly infection routes nicely
}
