% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_simulation_process.R
\name{plot_prev}
\alias{plot_prev}
\title{Plot the change in prevalence}
\usage{
plot_prev(
  param,
  output_filename = param$output_filename,
  output_dir = param$output_dir,
  i_simulation = 1:param$n_simulation,
  list_cows = NULL,
  language = NULL,
  title = T,
  xlab = T,
  ylab = T,
  font = NULL
)
}
\arguments{
\item{param, output_filename, output_dir}{See \link{param}.}

\item{i_simulation}{csvs with this numbers are used.}

\item{list_cows}{List consisted of \code{cow_table}s. Specify one of \code{output_dir}+\code{output_filename} or \code{list_cows}.}

\item{language}{Language to which translate messages. At present, only English and Japanese is implemented.}

\item{title, xlab, ylab}{logical or character. Plot title, label for x-axis and label for y-axis. When \code{TRUE}, the default value is used. When \code{FALSE}, a title is not shown (\code{TRUE} is valid only for \code{title}). When specified by character, the string is used as a title or label.}

\item{font}{Font in a plot. The default is "Meiryo" for Windows and "Hiragino Kaku Gothic Pro" for the other OS.}
}
\value{
An scatterplot by \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Plot monthly prevalences from csv files or list of \code{cow_table}. Set either one of \code{output_dir}+\code{output_filename} or \code{csv}.
}
