% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_simulation_process.R
\name{plot_prevalences}
\alias{plot_prevalences}
\title{Plot the change in prevalence}
\usage{
plot_prevalences(simulation_length, path_to_csv)
}
\arguments{
\item{simulation_length}{See \link{param_simulation}.}

\item{path_to_csv}{Path to a simulation output csv file.}
}
\value{
An scatterplot by \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Plot the change in prevalence
}
