% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_simulation_process.R
\name{plot_route}
\alias{plot_route}
\title{Plot monthly infection routes nicely}
\usage{
plot_route(
  param,
  output_filename = param$output_filename,
  output_dir = param$output_dir,
  i_simulation = seq_len(param$n_simulation),
  list_cows = NULL,
  language = NULL,
  drop = T,
  route_levels = NULL,
  route_labels = NULL,
  max_ylim = NULL,
  title = T,
  legend_title = T,
  xlab = T,
  ylab = T,
  gray = F,
  area_color = NULL,
  border = F,
  border_color = NULL,
  font = NULL
)
}
\arguments{
\item{param, output_filename, output_dir}{See \link{param}.}

\item{i_simulation}{csvs with this numbers are used.}

\item{list_cows}{List consisted of \code{cow_table}s. Specify one of \code{output_dir}+\code{output_filename} or \code{list_cows}.}

\item{language}{Language to which translate messages. At present, only English and Japanese is implemented.}

\item{drop}{Drop infection routes not in \code{csv} or \code{cows} from a legend.}

\item{route_levels, route_labels}{See \link{redefine_route_levels}}

\item{max_ylim}{Upper limit of the y-axis of the plot.}

\item{title, legend_title, xlab, ylab}{logical or character. Plot a title, a legend title, a label for x-axis and a label for y-axis. When \code{TRUE}, the default value is used. When \code{FALSE} or \code{NULL}, the title or the label is not shown. When specified by character, the string is used as the title or the label.}

\item{gray}{When \code{TRUE}, a plot will be a grayscale image.}

\item{area_color}{Specify a color palette of a plot.}

\item{border}{When \code{TRUE}, each area in a plot will be surrounded by border.}

\item{border_color}{Specify a color palette for the border.}

\item{font}{Font in a plot. The default is "Meiryo" for Windows and "Hiragino Kaku Gothic Pro" for the other OS.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} plot.
}
\description{
Plot average (median) monthly prevalences by each infection route from csv files or list of \code{cow_table}. Set either one of \code{output_dir}+\code{output_filename} or \code{csv}.
}
