% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_simulation_process.R
\name{prepare_communal_pasture}
\alias{prepare_communal_pasture}
\title{Process raw communal pasture use data to suitable form}
\usage{
prepare_communal_pasture(
  csv,
  data = NULL,
  output_file = NULL,
  area_name = NULL,
  sep = "[,\\t\\r\\n |;:]"
)
}
\arguments{
\item{csv}{File path of an input csv file. See the Detail section to know about form of input csv.}

\item{data}{data.frame as a input instead of \code{csv}. See the Detail section to know about form of input data.}

\item{output_file}{The name of an output file (must be a csv file). If \code{NULL}, no output file is created.}

\item{area_name}{If \code{area_out} and \code{area_back} are specified by character, specify integer \code{area_id} like \code{c(barnA = 1, barnB = 2, ...)}.}

\item{sep}{Separatator used in \code{priority} column. See explanation of \code{priority} in Detail section.}
}
\value{
A csv file which can be used as an input for \code{\link[=simulate_BLV_spread]{simulate_BLV_spread()}}.
}
\description{
Transform an input csv into a suitable form, which is in a form of \link{communal_pasture_table}.
}
\details{
An input csv file can have following columns. The csv file must contain \code{area_out}, \code{area_back}, \code{condition_out} and \code{condition_back} column.
\itemize{
\item \code{area_out}
\item \code{area_back}: A character like \code{"1|2|3"} will be converted to a character vector \code{c(1, 2, 3)}. Separator (\code{|}) can be specifed by \code{sep} argument. (This transformation from character to numeric is necessary if you want to read data from a csv file.)
\item \code{condition_out}
\item \code{condition_back}
\item \code{priority}: If \code{NA}, \code{rep(1, length(area_back))} is set (which means all \code{area_back}s have the same priority and cows are randomly allocated among all \code{area_back}s). Values must be integer/numeric vectors of the same length with \code{area_back} or \code{NA}. Multiple values can be specified like \code{area_back}.
}

For further detail of each variable, see \link{communal_pasture_table}.
}
