% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_simulation_process.R
\name{prepare_data}
\alias{prepare_data}
\title{Process raw data to suitable forms}
\usage{
prepare_data(
  excel,
  param,
  output = F,
  cow_data = NULL,
  area_data = NULL,
  movement_data = NULL,
  communal_pasture_data = NULL,
  cow_output_file = NULL,
  area_output_file = NULL,
  movement_output_file = NULL,
  communal_pasture_output_file = NULL,
  sep = "[,\\t\\r\\n |;:]",
  ...
)
}
\arguments{
\item{excel}{Set this or \code{cow_data}, \code{area_data} and \code{movement_data}.}

\item{param}{See \link{param}.}

\item{output}{When \code{TRUE}, create output csv files with names of "cow.csv", "area.csv", "movement.csv" and "communal_pasture.csv" into a working directory. Shorthand form of setting "xxx.csv" to \code{xxx_output_file}.}

\item{cow_data}{See \code{\link[=prepare_cow]{prepare_cow()}} for detail.}

\item{area_data}{See \code{\link[=prepare_area]{prepare_area()}} for detail.}

\item{movement_data}{See \code{\link[=prepare_movement]{prepare_movement()}} for detail.}

\item{communal_pasture_data}{See \code{\link[=prepare_communal_pasture]{prepare_communal_pasture()}} for detail.}

\item{cow_output_file, area_output_file, movement_output_file, communal_pasture_output_file}{If not \code{NULL}, created data is exported to the files with these names (must be csv files).}

\item{sep}{Separatator used in \code{capacity} column of area data. See \code{\link[=prepare_area]{prepare_area()}} for detail.}

\item{...}{Other arguments passed to \code{\link[=prepare_cow]{prepare_cow()}}.}
}
\value{
csv files which can be used as an input for \code{\link[=simulate_BLV_spread]{simulate_BLV_spread()}}.
}
\description{
Process raw data to suitable forms to use in simulation.
}
