% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{process_param}
\alias{process_param}
\title{Calculate parameters based on other parameters}
\usage{
process_param(setup_cows_res, param_simulation, param_farm, param_area)
}
\arguments{
\item{setup_cows_res}{A result of \code{\link[=setup_cows]{setup_cows()}}.}

\item{param_simulation}{See \link{param_simulation}.}

\item{param_farm}{See \link{param_farm}.}

\item{param_area}{See \link{param_area}.}
}
\value{
A list of calculated parameters.
}
\description{
\itemize{
\item \code{is_ts}: Wheter barns are tie-stall.
\item \code{ts_area}: Area ID of tie-stall Areas.
\item \code{is_md_separated_in_ts}: Whether milking and dry cows are separated.
\item \code{capacity}: The result of \code{\link[=set_capacity]{set_capacity()}}.
\item \code{prob_rep}: The result of \code{\link[=set_prob_rep]{set_prob_rep()}}. The probability that a newborn female calf will be a replacement cow.
}
}
\details{
Parameters processed by \code{\link[=process_param]{process_param()}} are deteministic. Parameters calculated by \code{\link[=calc_param]{calc_param()}} are stochastic.
}
