% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_simulation_process.R
\name{process_raw_area}
\alias{process_raw_area}
\title{Process raw area data to suitable form}
\usage{
process_raw_area(csv, data = NULL, output_file = NULL,
  sep = "[,\\t\\r\\n |;:]")
}
\arguments{
\item{csv}{File path of an input csv file. See the Detail section to know about form of input csv.}

\item{data}{data.frame as a input instead of \code{csv}. See the Detail section to know about form of input data.}

\item{output_file}{The name of an output file (must be a csv file). If \code{NULL}, no output file is created.}

\item{sep}{Separatator used in \code{capacity} column. See explanation of \code{capacity} in Detail section.}
}
\value{
A csv file which can be used as an input for \code{\link[=simulate_BLV_spread]{simulate_BLV_spread()}}.
}
\description{
Transform an input csv into a suitable form, which is in a form of \link{area_table}.
}
\details{
An input csv file can have following columns. The csv file must contain \code{area_type} column.
\itemize{
\item \code{area_id}: If not set or non-numerical value is set, sequencial integers are allocated (from 1 to the number of input rows). More than two rows can have the same \code{area_id} only when these rows have \code{area_type}s as "tie". \emph{e.g.} \code{data.frame(area_id = c(1, 1), area_type = c("tie", "tie"), capacity = c(10, 20))} is identical to \code{data.frame(area_id = 1, area_type = "tie", capacity = list(c(10, 20)))}. If \code{NA}, the previous non-\code{NA} value is set.
\item \code{area_type}
\item \code{capacity}: If \code{NA}, \code{Inf} is set. If \code{area_type} is "free", \code{capacity} must be set. A character like \code{"1|2|3"} will be converted to a numeric vector \code{c(1, 2, 4)}. Separator (\code{|}) can be specifed by \code{sep} argument. (This transformation from character to numeric is necessary if you want to read data of a farm having a tie-stall barn from a csv file.)
}

For further detail of each variable, see \link{area_table}.
}
