% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_simulation_process.R
\name{process_raw_csv}
\alias{process_raw_csv}
\title{Process raw cow data to suitable form}
\usage{
process_raw_csv(csv, data = NULL, today = Sys.Date(),
  create_calf_data = F, modify_prevalence = NULL,
  param_calculated = calc_param(param_farm, param_simulation),
  n_chambers = NULL)
}
\arguments{
\item{csv}{File path of input csv file. See the Detail section to know about form of input csv.}

\item{data}{data.frame as a input instead of \code{csv}. See the Detail section to know about form of input data.}

\item{today}{A Date class object or a character in "YYYY/MM/DD" format. The date used to calculate \code{age} from \code{date_birth} when \code{age} is not set. \code{today} is automatically calculated when both of \code{age} and \code{date_birth} are filled and \code{date_birth} is in form of Date rather than number (which means that the cow was born $n$ month ago) and the value passed to this argument is ignored.}

\item{create_calf_data}{logical or a numeric. Create data for young cows based on cow data in the input. Set this argument when the input does not contain data for young cows (e.g. when you use Nyuken data). If \code{TRUE}, create cows younger than the youngest cows in the input. If a numeric is set, create cows equal to or younger than that age.}

\item{modify_prevalence}{double (0-1). If not \code{NULL}, modify \code{infection_status} column to make prevalence to the specified value.}

\item{param_calculated}{The result from \link{calc_param}.}

\item{n_chambers}{Set if a farm owns tie-stall barns. Specify the number of chambers in each tie-stall barn like \code{c(area_id = the_number_of_chambers_in_the_area, ...)}.}
}
\value{
A csv file which can be used as an input for \code{\link[=simulate_BLV_spread]{simulate_BLV_spread()}}.
}
\description{
Transform an input csv into a suitable form, which is in a form of \link{cow_table}.
}
\details{
An input csv file can have following columns. The csv file must contain \code{age} column.
\itemize{
\item \code{cow_id}: It can be use-defined cow ID (not have to be a 9- or 10-digits code). If not set, \code{cow_id} is allocated sequencially (From 1 to the number of input cows).
\item \code{age}: Age in month. Either one of \code{age} or \code{date_birth} must be set.
\item \code{date_birth}: Birth day. If \code{age} is not set, \code{age} is calculated from this and \code{today} argument in the function.
\item \code{sex}: One of "female", "male" or "freemartin". If not set, all cows are assumed to be "female".
\item \code{is_replacement}: If a farm keeps non-replacement cows (e.g. a male newborn which will be send to a livestock market), set this parameter. If not set, \code{age} = 0 female cows are assumed to be a replacement according to \code{prop_replacement} parameter in \code{param_farm} and all male cows are assumed to be a non-replacement.
\item \code{stage}, \code{parity}, \code{date_last_delivery}, \code{date_got_pregnant}, \code{date_dried}: If not set, they will be calculated in according to parameters related with reproduction in \link{param_farm}.
\item \code{is_to_test_pregnancy}: If not set, \code{FALSE} is set.
\item \code{n_ai}: If not set, it is assumed to be 0.
\item \code{infection_status}: At least one of this variable or \code{modify_prevalence} argument must be set. Valid categories are follows: "al", "pl" and "ebl" (case insensitive). Other values or \code{NA} will be coerced to "s" (= non-infected). When \code{modify_prevalence} is set, prevalence is modified to make prevalence equal to the value of \code{modify_prevalence}.
\item \code{date_ial}, \code{date_ipl}, \code{date_ebl}: Specify the date when infection status was confirmed. If \code{NULL}, \code{0} is set.
\item \code{area_id}: If not set, cows are divided to four areas based on \code{stage}. If \code{NA}s are included, cows are allocated to areas in which cows with the same stage and parity are kept.
\item \code{month_in_area}: If not set, it is assumed to be 0. This parameter has no effect when a farm does not use \code{month_in_area} as a criteria for area movement. See \link{area_table} for detail of area movement.
\item \code{chamber_id}: If not set, it is randomly allocated later in \code{\link[=setup_cows]{setup_cows()}}.
\item \code{is_isolated}: If not set, \code{FALSE} is set.
}

For further detail of each variables, see \link{cow_table}.
}
