% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_simulation_process.R
\name{process_raw_movement}
\alias{process_raw_movement}
\title{Process raw movement data to suitable form}
\usage{
process_raw_movement(csv, data = NULL, output_file = NULL,
  area_name = NULL, sep = "[,\\t\\r\\n |;:]")
}
\arguments{
\item{csv}{File path of an input csv file. See the Detail section to know about form of input csv.}

\item{data}{data.frame as a input instead of \code{csv}. See the Detail section to know about form of input data.}

\item{output_file}{The name of an output file (must be a csv file). If \code{NULL}, no output file is created.}

\item{area_name}{If \code{current_area} and \code{next_area} are specified by character, specify integer \code{area_id} like \code{c(barnA = 1, barnB = 2, ...)}.}

\item{sep}{Separatator used in \code{priority} column. See explanation of \code{priority} in Detail section.}
}
\value{
A csv file which can be used as an input for \code{\link[=simulate_BLV_spread]{simulate_BLV_spread()}}.
}
\description{
Transform an input csv into a suitable form, which is in a form of \link{movement_table}.
}
\details{
An input csv file can have following columns. The csv file must contain \code{current_area}, \code{condition} and \code{next_area} column.
\itemize{
\item \code{current_area}
\item \code{condition}
\item \code{next_area}
\item \code{priority}: If \code{NA}, \code{rep(1, length(next_area))} is set (which means all \code{next_area}s have the same priority and cows are randomly allocated among all \code{next_area}s).
}

For further detail of each variable, see \link{movement_table}.
}
