% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{q975}
\alias{q975}
\title{97.5\% quantile of normal distribution}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
q975
}
\description{
97.5\% quantile of normal distribution
}
\keyword{datasets}
