% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_construct_filepath}
\alias{rcpp_construct_filepath}
\title{Construct filepath to output a file}
\usage{
rcpp_construct_filepath(filename, i = NULL, subdir = NULL,
  ext = ".csv")
}
\arguments{
\item{filename}{character. アウトプットファイル名.}

\item{i}{double. The file name will be like filename01.csv.}

\item{subdir}{character. A file is saved into data/output/subdir.}

\item{ext}{character. Extension of the output file.}
}
\value{
character. data/output/subdir/filenameiext. eg. data/output/your_sub_directory/filename01.csv
}
\description{
Construct filepath to output a file
}
