% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_set_i_month}
\alias{rcpp_set_i_month}
\title{Add the month indicator}
\usage{
rcpp_set_i_month(cows, i)
}
\arguments{
\item{cows}{cow data table}

\item{i}{numeric. The index of month.}
}
\description{
Add the month indicator
}
