% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_simulation_process.R
\name{read_final_cows}
\alias{read_final_cows}
\title{Read cow information at the end of a simulation}
\usage{
read_final_cows(output_filename, output_dir, n_simulation,
  simulation_length, levels_route = NULL, labels_route = NULL)
}
\arguments{
\item{output_filename, output_dir, n_simulation, simulation_length}{See \link{param_simulation}.}

\item{levels_route, labels_route}{See \link{redefine_levels_route}.}
}
\value{
A \link{cow_table} with an additional column \code{i_simulation}.
}
\description{
Read information of cows which owned by a farm at the end of simulations from csv files and redefine infection routes.
}
\seealso{
\link{read_initial_cows}
}
