% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_simulation_process.R
\name{read_initial_cows}
\alias{read_initial_cows}
\title{Read cow_table at the start of a simulation from a csv and modify data ready to plot}
\usage{
read_initial_cows(path_to_csv, levels_route = NULL,
  labels_route = NULL)
}
\arguments{
\item{path_to_csv}{Path to a csv file which contains \link{cow_table}.}

\item{levels_route, labels_route}{See \link{redefine_levels_route}.}
}
\value{
A \link{cow_table} with an additional column \code{i_simulation}.
}
\description{
Read cow_table from a csv file, extract owned cows, set \code{i_simulation} column to 1, and redefine infection routes.
}
\seealso{
\link{read_final_cows}
}
