% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{read_param}
\alias{read_param}
\title{Read parameters from a txt file}
\usage{
read_param(path)
}
\arguments{
\item{path}{A path of a txt file which contains parameter information.}
}
\description{
Read parameters from a txt file which contains parameter information (param_simulationXXX.txt) and return a list consisted of the parameter values in the file.
}
