% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_simulation_process.R
\name{redefine_levels_route}
\alias{redefine_levels_route}
\title{Redefine infection routes}
\usage{
redefine_levels_route(cows, levels_route = NULL, labels_route = NULL)
}
\arguments{
\item{cows}{See \link{cow_table}.}

\item{levels_route}{If specified, infection routes not specified in \code{levels_route} are coarced into "other" category. See \code{cause_infection} in \link{cow_table} to know about default categories.}

\item{labels_route}{Specify if you want to rename categories.}
}
\value{
A \link{cow_table} with recategorized \code{cause_infection}.
}
\description{
Recategorize \code{cause_infection} column in a \code{cow_table}.
}
