% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_simulation_process.R
\name{redefine_route_levels}
\alias{redefine_route_levels}
\title{Redefine infection routes}
\usage{
redefine_route_levels(
  cows,
  language = NULL,
  route_levels = NULL,
  route_labels = NULL
)
}
\arguments{
\item{cows}{See \link{cow_table}.}

\item{route_levels}{If specified, infection routes not specified in \code{route_levels} are coarced into "other" category. See \code{cause_infection} in \link{cow_table} to know about default categories.}

\item{route_labels}{Specify if you want to rename categories.}

\item{to}{Language to which translate messages. At present, only English and Japanese is implemented.}
}
\value{
A \link{cow_table} with recategorized \code{cause_infection}.
}
\description{
Recategorize \code{cause_infection} column in a \code{cow_table}.
}
