% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barn.R
\name{remove_cows}
\alias{remove_cows}
\title{Remove cows from a herd}
\usage{
remove_cows(cows, areas, i, area_table, removed_row, cause_removal)
}
\arguments{
\item{cows}{See \link{cow_table}.}

\item{areas}{See \link{setup_areas}.}

\item{i}{The number of months from the start of the simulation.}

\item{area_table}{See \link{area_table}.}

\item{removed_row}{Row indice in \code{cow_table} of cows to be removed from current areas.}

\item{cause_removal}{A cause of removal of a cow.}
}
\value{
A \link{cow_table} in which \code{area_id} and \code{chamber_id} of specified cows are set as \code{NA}.
}
\description{
Assign \code{NA}s to \code{area_id} and \code{chamber_id} of specified cows.
}
