% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{replace_selected_cows}
\alias{replace_selected_cows}
\title{Replace selected cows}
\usage{
replace_selected_cows(cows, areas, area_table, cow_id_to_cull, i)
}
\arguments{
\item{cows}{See \link{cow_table}.}

\item{areas}{See \link{tie_stall_table}.}

\item{area_table}{See \link{area_table}.}

\item{cow_id_to_cull}{\code{cow_id} to remove from \code{cows}.}

\item{i}{The number of months from the start of the simulation.}
}
\value{
A \link{cow_table}.
}
\description{
Replace selected cows
}
