% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{resample}
\alias{resample}
\title{sample() from x rather than 1:x}
\usage{
resample(x, ...)
}
\arguments{
\item{x, ...}{See \code{\link[=sample]{sample()}}.}
}
\description{
Modified version of \code{\link[=sample]{sample()}} to sample from \code{x} rather than \code{1:x} when \code{x} has length 1 and is numeric.
}
\details{
See "sample()'s surprise" and "safer version" section in the example section in the help of \code{\link[=sample]{sample()}} to understand why this function is necessary.
}
\examples{
sample(9)  # length 9
blvibmjp:::resample(9)  # length 1
}
