% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{save_to_csv}
\alias{save_to_csv}
\alias{save_params_txt}
\alias{construct_filepath}
\title{Export a cow_table or the list of parameters to a file.}
\usage{
save_to_csv(cows, filename, i, subdir = NA)

save_params_txt(params, filename, i, subdir)

construct_filepath(filename, i = NULL, subdir = NA, ext = ".csv")
}
\arguments{
\item{cows}{See \link{cow_table}.}

\item{filename}{The file name to an output file.}

\item{i}{The number of months from the start of the simulation.}

\item{subdir}{Specify the path to the directory from "output" folder if you want to output a file into a different directory.}

\item{params}{The parameter list used in the simulation. For more detail, see \code{\link[=calc_params]{calc_params()}}.}

\item{ext}{The file extension of an output file.}
}
\value{
An output file at the specified location (\code{save_to_csv()} and \code{save_params()}) or a character string indicates a file path (eg. "subdir/filename001.csv") (\code{construct_filepath()}).
}
\description{
Export a cow_table or the list of parameters to a file.
}
