% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barns.R
\name{set_capacity}
\alias{set_capacity}
\title{Calculate capacity of a barn based on inputed parameters}
\usage{
set_capacity(herd_size)
}
\arguments{
\item{herd_size}{The herd size in a simulated herd.}
}
\value{
numeric
}
\description{
Calculate capacity of a barn based on inputed parameters
}
