% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{set_i_month}
\alias{set_i_month}
\title{Set the variable i_month in a cow_table}
\usage{
set_i_month(cows, i)
}
\arguments{
\item{cows}{See [cow_table].}

\item{i}{The number of months from the start of the simulation.}
}
\value{
A [cow_table]
}
\description{
Set the variable i_month in a cow_table
}
