% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{set_param}
\alias{set_param}
\title{Overwrite default parameters with herd specific parameters}
\usage{
set_param(parameter, default)
}
\arguments{
\item{parameter}{The farm specific parameter}

\item{default}{The default parameter}
}
\value{
A list of overwritten parameters
}
\description{
It's used to overwrite default parameters (averages of Hokkaido or Japan) with farm specific parameters.  
When 'parameter' is not NA (= when a farm specific parameter was set), it returns the farm specific parameters.  
When 'parameter' is NA (= when a farm specific parameter was not set), it returns dafault parameters.
}
\note{
This function is used to find this environment (= the environment test_params defined). Be careful to use the function in the other file.
}
