% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_area_table}
\alias{setup_area_table}
\title{Setup of \code{area_table}}
\usage{
setup_area_table(area_table)
}
\arguments{
\item{area_table}{See \link{area_table}.}
}
\description{
Make \code{area_table}.
}
\seealso{
\link{area_table} \link{setup_cows} \link{setup_rp_table} \link{setup_areas}
}
