% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_areas}
\alias{setup_areas}
\title{Setup of \code{tiestall_table}}
\usage{
setup_areas(init_cows, param_area)
}
\arguments{
\item{init_cows}{The element \code{init_cows} of a result of \code{\link[=setup_cows]{setup_cows()}}.}

\item{param_area}{See \link{param_area}.}
}
\value{
A \link{tiestall_table}.
}
\description{
Make chamber matrix, which indicates in which chamber each cow is.
Cows kept in free-stall or paddock are not shown in this matrix.
}
\seealso{
\link{setup_rp_table} \link{tiestall_table} \link{setup_cows} \link{setup_area_table}
}
