% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_rp_table}
\alias{setup_rp_table}
\title{Setup of \code{rp_table}}
\usage{
setup_rp_table(param)
}
\arguments{
\item{param}{See \link{param}.}
}
\description{
Make initial \code{rp_table}.
}
\seealso{
\link{setup_cows} \link{setup_area_table} \link{rp_table} \link{setup_area_table}
}
