% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_function.R
\name{sex_newborns}
\alias{sex_newborns}
\alias{sex_twins}
\title{Sex ratio of newborns}
\usage{
sex_newborns(n_newborns, param_sim)

sex_twins(n_calves, param_sim)
}
\arguments{
\item{n_newborns, n_calves}{The number of newborns.}

\item{param_sim}{A list which combined \link{param}, a result of \code{\link[=process_param]{process_param()}} and a result of \code{\link[=calc_param]{calc_param()}}.}
}
\value{
A character vector consisted of "male", "female" and "freemartin".
}
\description{
\code{sex_newborns()} returns sex of singleton newborns.
\code{sex_twins()} returns sex of pairs of twin newborns.
Probability to be triplets or more is ignored (=0).
}
