% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{susceptibility}
\alias{susceptibility}
\title{Calculate susceptibility of newborns to the pathogen}
\usage{
susceptibility(n_newborns, susceptibility_ial_to_ipl_dam,
  susceptibility_ipl_to_ebl_dam)
}
\arguments{
\item{n_newborns}{The number of newborns}

\item{susceptibility_ial_to_ipl_dam, susceptibility_ipl_to_ebl_dam}{The dam's susceptibility to the pathogen}
}
\value{
A list consisted of susceptibility of newborns
}
\description{
Calculate susceptibility of newborns to the pathogen
}
