% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_simulation_process.R
\name{table_infection_status}
\alias{table_infection_status}
\alias{summary_infection_status}
\title{Summarize infection status}
\usage{
table_infection_status(
  output_filename,
  output_dir,
  n_simulation,
  simulation_length
)

summary_infection_status(cows)
}
\arguments{
\item{output_filename}{See \link{param}.}

\item{output_dir}{See \link{param}.}

\item{n_simulation}{See \link{param}.}

\item{simulation_length}{See \link{param}.}

\item{cows}{A \code{cow_table} read by \link{read_final_cows}.}
}
\description{
Calculate monthly infection status at the end of simulations.
}
\seealso{
\link{table_route}
}
