% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_simulation_process.R
\name{table_status}
\alias{table_status}
\alias{summary_status}
\title{Summarize infection status}
\usage{
table_status(
  param,
  route_levels = NULL,
  route_labels = NULL,
  output_filename = param$output_filename,
  output_dir = param$output_dir,
  i_simulation = seq_len(param$n_simulation),
  drop
)

summary_status(cows)
}
\arguments{
\item{param}{See \link{param}.}

\item{route_levels}{See \link{redefine_route_levels}.}

\item{route_labels}{See \link{redefine_route_levels}.}

\item{output_filename}{See \link{param}.}

\item{output_dir}{See \link{param}.}

\item{i_simulation}{csvs with this numbers are read.}

\item{drop}{Drop infection routes not in \code{csv} or \code{cows} from a legend.}

\item{cows}{A result of \code{\link[=read_final_cows]{read_final_cows()}}.}
}
\description{
Calculate monthly infection status at the end of simulations.
}
\seealso{
\link{table_route}
}
