% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\docType{data}
\name{tie_stall_table}
\alias{tie_stall_table}
\alias{a_chamber}
\title{A data.table to store status of a tie-stall barn}
\format{\link[data.table:data.table]{data.table}}
\usage{
a_chamber
}
\description{
\code{tie_stall_table} is a \link[data.table:data.table]{data.table} to store status of tie-stall barns.
Each tie-stall barn have each \code{tie_stall_table}.
The rows are consisted of \code{a_chamber}, which indicates one chamber in a barn.
}
\details{
\code{chamber_id} and \code{adjoin_previous/next_chamber} are fixed. Values will not be changed while a simulation.
Other variables are flexible. Values will may be changed while a simulation.
\itemize{
\item \code{chamber_id}: Chamber ID.
\item \code{adjoin_previous_chamber}: Whether the chamber adjoins the \code{chamber_id - 1}th chamber.
\item \code{adjoin_next_chamber}: Whether the chamber adjoins the \code{chamber_id + 1}th chamber.
\item \code{cow_id}: Cow ID in a lane.
\item \code{cow_status}: Infection status of the cow.
\item \code{is_isolated}: Whether the cow is isolated or not.
}
}
\seealso{
\link{cow_table} \link{area_table} \link{movement_table} \link{rp_table}
}
\keyword{datasets}
