% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_simulation_process.R
\name{translate_msg}
\alias{translate_msg}
\title{Translate plot title and labels}
\usage{
translate_msg(type, to)
}
\arguments{
\item{type}{Type of massages.}

\item{to}{Language to which translate messages. At present, only English and Japanese is implemented.}
}
\description{
Translate plot title and labels to other languages than English.
At present, translation only for Japanese is implemented.
}
