% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{unassign_removed_cows}
\alias{unassign_removed_cows}
\title{Remove dead/culled/sold cows from \code{areas}}
\usage{
unassign_removed_cows(cows, area_table, areas, i)
}
\arguments{
\item{cows}{See \link{cow_table}.}

\item{area_table}{See \link{area_table}.}

\item{areas}{See \link{setup_areas} and \link{tie_stall_table}.}

\item{i}{The number of months from the start of the simulation.}
}
\description{
Remove dead/culled/sold cows from \code{areas}
}
