% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{update_if_input_not_null}
\alias{update_if_input_not_null}
\title{Update a list if an input is not \code{NULL}}
\usage{
update_if_input_not_null(list, ...)
}
\arguments{
\item{list}{A list to be updated.}

\item{...}{Inputs to update.}
}
\value{
A list whose components which has same names with non-\code{NULL} inputs are overwritten.
}
\description{
Overwrite components of a list if inputs are not \code{NULL}.
}
\examples{
a <- "A"
b <- NULL
l <- list(a = "a", b = "B", c = 3)
update_if_input_not_null(l, a, b)  # list(a = "A", b = "B", c = 3)
}
