% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_category}
\alias{validate_category}
\title{Validate that categorical columns contain valid categories only}
\usage{
validate_category(cows, col, category)
}
\arguments{
\item{cows}{A \link{cow_table}.}

\item{col}{Name of a column to validate.}

\item{category}{Possible categories.}
}
\value{
Stop if the \code{col} contains invalid category.
}
\description{
Validate that categorical columns contain valid categories only
}
