% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_cow_table}
\alias{validate_cow_table}
\title{Validate cow_table}
\usage{
validate_cow_table(cows, date_format = "ymd",
  today = lubridate::today(tzone = "Asia/Tokyo"))
}
\arguments{
\item{cows}{A \link{cow_table}.}

\item{date_format}{One of "ymd", "ydm", "myd", "mdy", "dym", "dmy".}

\item{today}{A \code{date} class object indicates today.}
}
\value{
A \link{cow_table} whose columns are fomatted.
}
\description{
Validate \code{cow_table}.
}
