% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{validate_param}
\alias{validate_param}
\title{Validate parameters}
\usage{
validate_param(param, list_param_modif = NULL)
}
\arguments{
\item{param}{See \link{param}.}

\item{list_param_modif}{See \code{\link[=simulate_blv_spread]{simulate_blv_spread()}}.}
}
\value{
A parameter list.
}
\description{
Validate parameters
}
