% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{vars_to_named_list}
\alias{vars_to_named_list}
\title{Make a named list}
\usage{
vars_to_named_list(...)
}
\arguments{
\item{...}{Inputs.}
}
\value{
A list consisted of inputs and having same names with inputs.
}
\description{
Construct a named list whose elements' names are same with the inputs.
}
\examples{
a <- "a"
b <- "B"
c <- 3
vars_to_named_list(a, b, c)  # list(a = "a", b = "B", c = 3)
}
