% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_simulation_process.R
\name{warn_double_sep}
\alias{warn_double_sep}
\title{Check whether input contains doublewidth separator}
\usage{
warn_double_sep(data)
}
\arguments{
\item{data}{Data to check like \code{movement_table$next_area}.}
}
\description{
Check whether input contains doublewidth separator
}
