% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_simulation_process.R
\name{warn_invalid_col}
\alias{warn_invalid_col}
\title{Check whether input contains invalid column(s)}
\usage{
warn_invalid_col(input, default_col)
}
\arguments{
\item{input}{A data.table}

\item{default_col}{A character vector indicating valid column names}
}
\description{
Check whether input contains invalid column(s)
}
