% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{add_1_to_age}
\alias{add_1_to_age}
\title{Increment age in a cow table}
\usage{
add_1_to_age(cows)
}
\arguments{
\item{cows}{See [cow_table].}
}
\value{
A [cow_table]
}
\description{
Add 1 to `age` variable in a cow table.
}
