% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{add_newborns}
\alias{add_newborns}
\title{Add newborns to a cow_table}
\usage{
add_newborns(cows, params_calculated, i)
}
\arguments{
\item{cows}{See [cow_table].}

\item{params_calculated}{See [calculate_parameters].}

\item{i}{The number of months from the start of the simulation.}
}
\value{
A [cow_table]
}
\description{
Add newborns to a cow_table
}
