% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{add_newborns}
\alias{add_newborns}
\title{Add newborns to a cow_table}
\usage{
add_newborns(cows, i, last_cow_id, param_calculated, param_processed)
}
\arguments{
\item{cows}{See \link{cow_table}.}

\item{i}{The number of months from the start of the simulation.}

\item{last_cow_id}{The ID of a cow at the last non-empty row of \code{cows}.}

\item{param_calculated}{Return from \code{\link[=calc_param]{calc_param()}}.}

\item{param_processed}{Return from \code{\link[=process_param]{process_param()}}.}
}
\value{
A list consisted of two elements: \code{cows} and \code{last_cow_id}.
}
\description{
Add newborns to a cow_table
}
