% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\docType{data}
\name{area_table}
\alias{area_table}
\alias{a_area}
\alias{area_table_cols}
\title{A data.table to manage areas in a farm}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1 rows and 3 columns.

An object of class \code{character} of length 3.
}
\usage{
a_area

area_table_cols
}
\description{
\code{area_table} is a \link[data.table:data.table]{data.table} to manage areas (barns, paddocks, hatches, etc.) in a farm.
Users must specify one \code{area_table} consisted of following items before starting a simulation.
}
\details{
\itemize{
\item \code{area_id} (integer): Area ID.
\item \code{area_type} (\code{"free"}/\code{"tie"}/\code{"outside"}/\code{"hatch"}/\code{"communal pasture"}): Type of a area. Specify one of \code{"free"} (hatch, freebarn, free-stall, etc.), \code{"tie"} (tie-stall), \code{"outside"} (paddock or rangeland, etc.), \code{"hatch"} or \code{"communal pasture"} (yotaku).
\item \code{capacity} (list consisted of numeric): Max number of cows to be kept in the area. \code{Inf} is set if you specify \code{NA}. \code{capacity} must be set if \code{area_type} is \code{"tie"}; otherwise optional.
\itemize{
\item If \code{area_type} is \code{"free"} or \code{"outside"}: a numeric or \code{NA}.
\item If \code{area_type} is \code{"tie"}: a numeric vector whose length is equal to the number of lanes in the area and each elements indicates the number of chambers in a lane.
\item If \code{area_type} is \code{"hatch"}: only \code{NA} is allowed.
}
}
}
\note{
Several parameters are calculated by \link{setup_area_table} and added to a \code{area_table} as attribute variables. Such values are intenended to be touched only by simulation functions and not by users.
\itemize{
\item \code{capacity}: Max number of cows can be kept in a area.
\item \code{tie_stall}: \code{area_id}s of tie-stall areas.
}
}
\examples{
# A farm has three areas:
# - A freebarn for calves (max capacity is 30 calves).
# - A paddock for heifers (capacity is not limited).
# - A tie-stall for delivered cows consisted of 2 lanes with 40 chambers and 2 lanes with 30 chambers.
areas <- a_area[rep(1, 3), ]
areas[, `:=`(area_id = 1:3,
             area_type = c("free", "outside", "tie"),
             capacity = list(30, NA, c(40, 40, 30, 30))]
}
\seealso{
\link{cow_table} \link{tie_stall_table} \link{movement_table} \link{rp_table}
}
\keyword{datasets}
