% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barn.R
\name{assign_chambers}
\alias{assign_chambers}
\title{Assign chamber_id to cows allocated to tie-stall barns}
\usage{
assign_chambers(cows, area_list, area_assignment)
}
\arguments{
\item{cows}{See \link{cow_table}.}

\item{area_list}{See \link{setup_areas}.}

\item{area_assignment}{See \code{\link[=calculate_area_assignment]{calculate_area_assignment()}}.}
}
\value{
A \link{cow_table}.
}
\description{
Assign \code{chamber_id} to cows allocated to tie-stall barns.
}
