% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barn.R
\name{calc_infection_in_barns}
\alias{calc_infection_in_barns}
\title{Calculate infection in barns}
\usage{
calc_infection_in_barns(cows, area_table, area_list)
}
\arguments{
\item{cows}{See \link{cow_table}.}

\item{area_table}{See \link{area_table}.}

\item{area_list}{See \link{setup_areas} and \link{tie_stall_table}.}
}
\value{
A \link{cow_table}.
}
\description{
Calculate infection in barns depending on barn type (tied or freed)
}
