% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{calc_param_both}
\alias{calc_param_both}
\title{Calculate parameters used in both of process_raw_data() and simulation}
\usage{
calc_param_both(param_simulation)
}
\value{
A parameter list.
}
\description{
Calculate parameters which are in both of \code{\link[=process_raw_data]{process_raw_data()}} and simulation
}
\seealso{
\code{\link[=calc_param]{calc_param()}} \code{\link[=calc_param_pre]{calc_param_pre()}}
}
