% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{change_area}
\alias{change_area}
\title{Check and move cows between areas}
\usage{
change_area(cows, movement_table, area_table, area_list, param_area, i)
}
\arguments{
\item{cows}{See \link{cow_table}.}

\item{movement_table}{See \link{movement_table}.}

\item{area_table}{See \link{area_table}.}

\item{area_list}{See \link{setup_areas} and \link{tie_stall_table}.}

\item{param_area}{See \link{param_area}.}

\item{i}{The number of months from the start of the simulation.}
}
\value{
A list composed of \link{cow_table} and \link{area_list}.
}
\description{
Check and move cows between areas
}
