% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{change_infection_status}
\alias{change_infection_status}
\title{Check change of infection status of cows}
\usage{
change_infection_status(cows, i, month)
}
\arguments{
\item{cows}{See [cow_table].}

\item{i}{The number of months from the start of the simulation.}

\item{month}{The month (1, 2, ..., 12) of month `i`.}
}
\value{
A [cow_table]
}
\description{
Check change of infection status of cows
}
