% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{change_stage}
\alias{change_stage}
\title{Change stage of cows and move cows between barns accordinglly}
\usage{
change_stage(cows, groups, i)
}
\arguments{
\item{cows}{See [cow_table].}

\item{groups}{See [barn_table].}

\item{i}{The number of months from the start of the simulation.}
}
\value{
A list consists of `cow_table` and `barn_table`
}
\description{
Change stage of cows and move cows between barns accordinglly
}
