% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{check_removal}
\alias{check_removal}
\title{Check death and sale of current cows}
\usage{
check_removal(cows, groups, i)
}
\arguments{
\item{cows}{See \link{cow_table}.}

\item{groups}{See \link{barn_table}.}

\item{i}{The number of months from the start of the simulation.}
}
\value{
A list consisted of \code{cow_table} and \code{barn_table}
}
\description{
Check death and sale of current cows
}
