% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{check_removal}
\alias{check_removal}
\title{Check death and sale of current cows}
\usage{
check_removal(cows, groups, i, param_calculated, param_processed)
}
\arguments{
\item{cows}{See \link{cow_table}.}

\item{groups}{See \link{tiestall_table}.}

\item{i}{The number of months from the start of the simulation.}

\item{param_calculated}{Return from \code{\link[=calc_param]{calc_param()}}.}

\item{param_processed}{Return from \code{\link[=process_param]{process_param()}}.}
}
\value{
A list consisted of \link{cow_table} and \link{tiestall_table}.
}
\description{
Check death and sale of current cows
}
