% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{check_removal}
\alias{check_removal}
\title{Check death and sale of current cows}
\usage{
check_removal(cows, areas, i, area_table, param_sim)
}
\arguments{
\item{cows}{See \link{cow_table}.}

\item{areas}{See \link{tie_stall_table}.}

\item{i}{The number of months from the start of the simulation.}

\item{param_sim}{A list which combined \link{param}, a result of \code{\link[=process_param]{process_param()}} and a result of \code{\link[=calc_param]{calc_param()}}.}
}
\value{
A list consisted of \link{cow_table} and \link{tie_stall_table}.
}
\description{
Check death and sale of current cows
}
