% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{cull_infected_cows}
\alias{cull_infected_cows}
\title{Cull infected cows}
\usage{
cull_infected_cows(cows, i, param_farm, param_calculated)
}
\arguments{
\item{cows}{See \link{cow_table}.}

\item{i}{The number of months from the start of the simulation.}

\item{param_farm}{See \link{param_farm}.}

\item{param_calculated}{Return from \code{\link[=calc_param]{calc_param()}}.}
}
\value{
A \link{cow_table}.
}
\description{
Cull infected cows
}
