% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_simulation_process.R
\name{define_msg}
\alias{define_msg}
\title{Define plot title and labels}
\usage{
define_msg(original_msg, default_msg, type, to)
}
\arguments{
\item{original_msg}{List of original title and labels before passed to \code{\link[=translate_msg]{translate_msg()}}.}

\item{default_msg}{List of default plot title and labels.}

\item{type}{Type of massages.}

\item{to}{Language to which translate messages. At present, only English and Japanese is implemented.}
}
\description{
Define plot title and labels based on arguments
}
