% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{do_ai}
\alias{do_ai}
\title{Conduct AI and check chance of infection}
\usage{
do_ai(cows, i, day_rp, param_calculated)
}
\arguments{
\item{cows}{See \link{cow_table}.}

\item{i}{The number of months from the start of the simulation.}

\item{day_rp}{See \link{rp_table}.}

\item{param_calculated}{Return from \code{\link[=calc_param]{calc_param()}}.}
}
\value{
A \link{cow_table}.
}
\description{
Conduct AI and check chance of infection
}
