% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{heat_cycle}
\alias{heat_cycle}
\title{Heat cycle}
\usage{
heat_cycle(n_cows)
}
\arguments{
\item{n_cows}{The number of cows to which heat cycle should be calculated}
}
\value{
A numeric vector
}
\description{
Heat cycle
}
