% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{is_ai_started_heifer}
\alias{is_ai_started_heifer}
\title{Whether the first AIs for hifers are conducted}
\usage{
is_ai_started_heifer(ages)
}
\arguments{
\item{ages}{Age of heifers}
}
\value{
A logical vector
}
\description{
Whether the first AIs for hifers are conducted
}
