% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{is_dried}
\alias{is_dried}
\title{Whether a cow is dried}
\usage{
is_dried(months_from_delivery)
}
\arguments{
\item{months_from_delivery}{The number of months past from the last delivery}
}
\value{
A logical vector
}
\description{
Whether a cow is dried
}
