% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_function.R
\name{is_ebl_detected}
\alias{is_ebl_detected}
\title{Whether a EBL cow is detected}
\usage{
is_ebl_detected(id_cow_ebl, param_sim)
}
\arguments{
\item{param_sim}{A list which combined \link{param}, a result of \code{\link[=process_param]{process_param()}} and a result of \code{\link[=calc_param]{calc_param()}}.}

\item{rows_cow_ebl}{The row indices in \link{cow_table} of EBL cows.}
}
\value{
A logical vector.
}
\description{
Whether a EBL cow is detected
}
