% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{is_ebl_detected}
\alias{is_ebl_detected}
\title{Whether a EBL cow is detected}
\usage{
is_ebl_detected(id_cow_ebl)
}
\arguments{
\item{id_cow_ebl}{The ID of EBL cows.}
}
\value{
A logical vector
}
\description{
Whether a EBL cow is detected
}
