% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{is_heat_detected}
\alias{is_heat_detected}
\title{Whether a heat is detected}
\usage{
is_heat_detected(n_cows)
}
\arguments{
\item{n_cows}{The number of cows which come into heat.}
}
\value{
A logical vector
}
\description{
Whether a heat is detected
}
