% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_function.R
\name{is_infected_communal_pasture}
\alias{is_infected_communal_pasture}
\title{Whether cows are infected in a communal pasture}
\usage{
is_infected_communal_pasture(n_cows, param_farm)
}
\arguments{
\item{n_cows}{The number of cows.}

\item{param_farm}{See \link{param_farm}.}
}
\value{
A logical vector.
}
\description{
Whether cows are infected in a communal pasture
}
