% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{is_infected_contact}
\alias{is_infected_contact}
\title{Whether cows are infected by direct contact}
\usage{
is_infected_contact()
}
\arguments{
\item{id_cow_s}{ID of infected cows.}
}
\value{
A logical vector
}
\description{
Whether cows are infected by direct contact
}
