% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_function.R
\name{is_infected_insects}
\alias{is_infected_insects}
\title{Whether cows are infected by insects}
\usage{
is_infected_insects(n_cows, month, param_calculated)
}
\arguments{
\item{n_cows}{The number of cows.}

\item{month}{The current month (1, 2, ..., 12).}

\item{param_calculated}{Return from \code{\link[=calc_param]{calc_param()}}.}
}
\value{
A logical vector.
}
\description{
Whether cows are infected by insects
}
