% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{is_infected_insects}
\alias{is_infected_insects}
\title{Wheter cows are infected by insects}
\usage{
is_infected_insects(id_cow_s, month)
}
\arguments{
\item{id_cow_s}{ID of infected cows.}

\item{month}{The current month (1, 2, ..., 12)}
}
\value{
A logical vector
}
\description{
Wheter cows are infected by insects
}
