% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_function.R
\name{is_infected_needles}
\alias{is_infected_needles}
\title{Whether cows are infected by contaminated needles}
\usage{
is_infected_needles(id_cow_s, param_calculated)
}
\arguments{
\item{id_cow_s}{\code{cow_id} of infected cows.}

\item{param_calculated}{Return from \code{\link[=calc_param]{calc_param()}}.}
}
\value{
A logical vector.
}
\description{
Whether cows are infected by contaminated needles
}
