% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_function.R
\name{is_infected_needles}
\alias{is_infected_needles}
\title{Whether cows are infected by contaminated needles}
\usage{
is_infected_needles(n_cows, cows, param_calculated)
}
\arguments{
\item{n_cows}{The number of cows.}

\item{cows}{See \link{cow_table}.}

\item{param_calculated}{Return from \code{\link[=calc_param]{calc_param()}}.}
}
\value{
A logical vector.
}
\description{
Whether cows are infected by contaminated needles
}
