% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_function.R
\name{is_infected_rp}
\alias{is_infected_rp}
\title{Whether cows are infected by rectal palpation}
\usage{
is_infected_rp(n_cows_palpated, param_sim)
}
\arguments{
\item{n_cows_palpated}{ID of rectally palpated cows.}

\item{param_sim}{A list which combined \link{param}, a result of \code{\link[=process_param]{process_param()}} and a result of \code{\link[=calc_param]{calc_param()}}.}
}
\value{
A logical vector.
}
\description{
Whether cows are infected by rectal palpation
}
