% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_function.R
\name{is_infected_rp}
\alias{is_infected_rp}
\title{Whether cows are infected by rectal palpation}
\usage{
is_infected_rp(n_cows_palpated, param_calculated)
}
\arguments{
\item{n_cows_palpated}{ID of rectally palpated cows.}

\item{param_calculated}{Return from \code{\link[=calc_param]{calc_param()}}.}
}
\value{
A logical vector.
}
\description{
Whether cows are infected by rectal palpation
}
