% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{is_infected_rp}
\alias{is_infected_rp}
\title{Whether cows are infected by rectal palpation}
\usage{
is_infected_rp(n_cows_palpated)
}
\arguments{
\item{n_cows_palpated}{IDs of rectally palpated cows}
}
\value{
A logical vector
}
\description{
Whether cows are infected by rectal palpation
}
