% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{is_infected_vertical}
\alias{is_infected_vertical}
\title{Wheter newborns are infected vertically}
\usage{
is_infected_vertical(status_mother)
}
\arguments{
\item{status_mother}{The `infection_status` of dams.}
}
\value{
A logical vector
}
\description{
Wheter newborns are infected vertically
}
