% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barn.R
\name{is_md_separated_in_ts}
\alias{is_md_separated_in_ts}
\title{Check whether milking cows and dry cows are separated}
\usage{
is_md_separated_in_ts(param_calculated)
}
\arguments{
\item{param_calculated}{A list of parameters. See \code{\link[=calc_param]{calc_param()}}.}
}
\value{
A logical value.
}
\description{
Check whether milking cows and dry cows are separated
}
