% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{set_prob_rep}
\alias{set_prob_rep}
\alias{is_replacement}
\title{Whether a newborn will be a replacement}
\usage{
set_prob_rep(n_delivered)

is_replacement(n_calves, capacity, herd_size)
}
\arguments{
\item{n_delivered}{The number of delivered cows in a herd.}

\item{n_calves}{The number of newborns.}

\item{capacity}{Capacity of a herd.}

\item{herd_size}{The current herd size.}
}
\value{
A logical vector or a numeric value (`set_prob_rep()`)
}
\description{
Whether a newborn will be a replacement
}
\references{
「平成28年度 乳用種初生牛の経営に関する調査報告書」の「表40 調査対象経営の乳用種雌子牛の仕向け状況（規模別）」より
}
