% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{is_stillbirth}
\alias{is_stillbirth}
\title{Whether a delivery end up in stillbirth or abortion}
\usage{
is_stillbirth(parity)
}
\arguments{
\item{parity}{Parity of dams}
}
\value{
A logical vector
}
\description{
Here, stillbirth means stillbirth and abortion
}
