% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_function.R
\name{is_stillbirth}
\alias{is_stillbirth}
\title{Whether a delivery end up in stillbirth or abortion}
\usage{
is_stillbirth(parity, param_calculated)
}
\arguments{
\item{parity}{Parity of dams.}

\item{param_calculated}{Return from \code{\link[=calc_param]{calc_param()}}.}
}
\value{
A logical vector.
}
\description{
Here, stillbirth means stillbirth and abortion.
}
