% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barn.R
\name{is_ts}
\alias{is_ts}
\title{Check wheter a area is a tie-stall barn}
\usage{
is_ts(param_area)
}
\arguments{
\item{param_area}{See \link{param_area}.}
}
\value{
A logical vector of length 4 which indicates whether areas are tie-stall or not.
}
\description{
Check wheter a area is a tie-stall barn
}
