% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barns.R
\name{is_ts}
\alias{is_ts}
\title{Check wheter a barn is tie-stall}
\usage{
is_ts(params)
}
\arguments{
\item{params}{A list of parameter. See \code{\link[=calc_params]{calc_params()}}.}
}
\value{
A logical vector of length 4 which indicates whether barns are tie-stall or not.
}
\description{
Check wheter a barn is tie-stall
}
