% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_function.R
\name{age_die}
\alias{age_die}
\alias{longevity}
\alias{age_slaughtered}
\title{Expected age of death or slaugher}
\usage{
age_die(n_cows, param_calculated)

age_slaughtered(n_cows, param_calculated)

longevity(n_cows, param_calculated)
}
\arguments{
\item{n_cows}{The number of cows to calculate expected age of death.}

\item{param_calculated}{Return from \code{\link[=calc_param]{calc_param()}}.}
}
\value{
A numeric vector.
}
\description{
Expected age of death or slaugher
}
