% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{age_die}
\alias{age_die}
\alias{age_slaughtered}
\alias{longevity}
\title{Expected age of death or slaugher}
\usage{
age_die(n_cows)

age_slaughtered(n_cows)

longevity(n_cows)
}
\arguments{
\item{n_cows}{The number of cows to calculate expected age of death}
}
\value{
A numeric vector
}
\description{
Expected age of death or slaugher
}
