% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_function.R
\name{n_newborn_per_dam}
\alias{n_newborn_per_dam}
\title{The number of newborns per dam}
\usage{
n_newborn_per_dam(n_dams, param_sim)
}
\arguments{
\item{n_dams}{The number of dams.}

\item{param_sim}{A list which combined \link{param}, a result of \code{\link[=process_param]{process_param()}} and a result of \code{\link[=calc_param]{calc_param()}}.}
}
\value{
A numeric vector consisted of 1 and 2.
}
\description{
The number of newborns per dam
}
