% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{n_newborn_per_dam}
\alias{n_newborn_per_dam}
\title{The number of newborns per dam}
\usage{
n_newborn_per_dam(n_dams)
}
\arguments{
\item{n_dams}{}
}
\value{
A numeric vector consisted of 1 and 2
}
\description{
The number of newborns per dam
}
