% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\docType{data}
\name{param_group}
\alias{param_group}
\title{Parameters about barns which should be set by users}
\format{An object of class \code{list} of length 4.}
\usage{
param_group
}
\description{
\itemize{
\item \code{n_group}: The number of barns.
\item \code{xy_chamber}: The number of chambers per lane and the number of lanes.
\item \code{is_calf_separated}: Whether each calf is separated.
\item \code{is_milking_dry_separated}: (For a farm which miling and dry cows are kept in same tie-stall barn,) whether dry cows are separated from milking cows.
}
}
\seealso{
\link{param_simulation} \link{param_farm}
}
\keyword{datasets}
