% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_simulation_process.R
\name{plot_infection_route}
\alias{plot_infection_route}
\title{Plot monthly infection routes nicely}
\usage{
plot_infection_route(path_to_csv, levels_route = NULL,
  labels_route = NULL, max_ylim = 100, title = NULL,
  legend_title = NULL, xlab = "Months in simulation",
  ylab = "Number of cattle", scale_fill = NULL)
}
\arguments{
\item{path_to_csv}{Path to an output csv file.}

\item{levels_route, labels_route}{See \link{redefine_levels_route}}

\item{max_ylim}{Upper limit of the y-axis of the plot.}

\item{title, legend_title, xlab, ylab}{Plot title, legend title, label for x-axis, label for y-axis.}

\item{scale_fill}{Specify a color palette of a plot.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} plot.
}
\description{
Plot monthly infection routes nicely
}
