% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_simulation_process.R
\name{plot_prev}
\alias{plot_prev}
\title{Plot the change in prevalence}
\usage{
plot_prev(
  simulation_length,
  csv,
  language = NULL,
  title = T,
  xlab = T,
  ylab = T,
  font = NULL
)
}
\arguments{
\item{simulation_length}{See \link{param}.}

\item{csv}{Path to a simulation output csv file.}

\item{language}{When set, plot title and labels are translated in this language. At present, only Japanese is implemented.}

\item{title, xlab, ylab}{logical or character. Plot title, label for x-axis and label for y-axis. When \code{TRUE}, the default value is used. When \code{FALSE}, a title is not shown (\code{TRUE} is valid only for \code{title}). When specified by character, the string is used as a title or label.}

\item{font}{Font in a plot. The default is "Meiryo" for Windows and "Hiragino Kaku Gothic Pro" for the other OS.}
}
\value{
An scatterplot by \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Plot the change in prevalence
}
