% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_simulation_process.R
\name{plot_prevalences}
\alias{plot_prevalences}
\title{Plot the change in prevalence}
\usage{
plot_prevalences(simulation_length, path_to_csv, language = NULL,
  title = "Change of prevalence", xlab = "Months in simulation",
  ylab = "Prevalence", font = NULL)
}
\arguments{
\item{simulation_length}{See \link{param_simulation}.}

\item{path_to_csv}{Path to a simulation output csv file.}

\item{language}{When set, plot title and labels are translated in this language. At present, only Japanese is implemented.}

\item{title, xlab, ylab}{Plot title, label for x-axis and label for y-axis.}

\item{font}{Font in a plot. The default is "Meiryo" for Windows and "Hiragino Kaku Gothic Pro" for the other OS.}
}
\value{
An scatterplot by \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Plot the change in prevalence
}
