% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_simulation_process.R
\name{process_raw_data}
\alias{process_raw_data}
\title{Process raw data to suitable forms}
\usage{
process_raw_data(cow_csv, area_csv, cow_data = NULL, area_data = NULL,
  output = F, cow_output_file = NULL, area_output_file = NULL,
  param_calculated = calc_param(param_farm, param_simulation),
  sep = "[,\\t\\r\\n |;:]")
}
\arguments{
\item{cow_csv, cow_data}{Set one of these arguments. See \code{\link[=process_raw_cow]{process_raw_cow()}} for detail.}

\item{area_csv, area_data}{Set one of these arguments. See \link{area_table} about detail of input.}

\item{output}{When \code{TRUE}, create output csv files with names of "cow.csv" and "area.csv" into a working directory. Shorthand form of setting "cow.csv" to \code{cow_output_file} and "area.csv" to \code{area_output_file}.}

\item{cow_output_file, area_output_file}{If not \code{NULL}, created data is exported to the file (must be csv file).}

\item{param_calculated}{The result from \link{calc_param}.}

\item{sep}{Separatator used in \code{capacity} column of area data. See \code{\link[=process_raw_area]{process_raw_area()}} for detail.}
}
\value{
csv files which can be used as an input for \code{\link[=simulate_BLV_spread]{simulate_BLV_spread()}}.
}
\description{
Process csv files to suitable forms to use in simulation.
}
