% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_simulation_process.R
\name{read_cows}
\alias{read_cows}
\title{Read cow information from simulation csvs}
\usage{
read_cows(
  param,
  route_levels = NULL,
  route_labels = NULL,
  output_filename = param$output_filename,
  output_dir = param$output_dir,
  n_simulation = param$n_simulation
)
}
\arguments{
\item{param, output_filename, output_dir, n_simulation}{See \link{param}.}

\item{route_levels, route_labels}{See \link{redefine_route_levels}.}
}
\value{
A \link{cow_table} with an additional column \code{i_simulation}.
}
\description{
Read information of cows which owned by a farm at the end of simulations from csv files and redefine infection routes.
}
