% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_simulation_process.R
\name{read_final_cows}
\alias{read_final_cows}
\title{Read cow information at the end of a simulation}
\usage{
read_final_cows(
  output_filename,
  output_dir,
  n_simulation,
  simulation_length,
  route_levels = NULL,
  route_labels = NULL
)
}
\arguments{
\item{output_filename, output_dir, n_simulation, simulation_length}{See \link{param}.}

\item{route_levels, route_labels}{See \link{redefine_route_levels}.}
}
\value{
A \link{cow_table} with an additional column \code{i_simulation}.
}
\description{
Read information of cows which owned by a farm at the end of simulations from csv files and redefine infection routes.
}
\seealso{
\link{read_initial_cows}
}
