% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_simulation_process.R
\name{read_initial_cows}
\alias{read_initial_cows}
\title{Read cow_table at the start of a simulation from a csv and modify data ready to plot}
\usage{
read_initial_cows(csv, route_levels = NULL, route_labels = NULL)
}
\arguments{
\item{csv}{Path to a csv file which contains \link{cow_table}.}

\item{route_levels, route_labels}{See \link{redefine_route_levels}.}
}
\value{
A \link{cow_table} with an additional column \code{i_simulation}.
}
\description{
Read cow_table from a csv file, extract owned cows, set \code{i_simulation} column to 0, and redefine infection routes.
}
\seealso{
\link{read_final_cows}
}
