% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barn.R
\name{remove_from_group}
\alias{remove_from_group}
\title{Remove dead or sold cows from a barn}
\usage{
remove_from_group(group, cow_id_removed)
}
\arguments{
\item{group}{See \link{barn_table}.}

\item{cow_id_removed}{The ID of cows removed from the barn.}
}
\value{
A \link{barn_table}.
}
\description{
Remove dead or sold cows from a barn
}
