% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barn.R
\name{set_capacity}
\alias{set_capacity}
\title{Calculate capacity of an area based on inputed parameters}
\usage{
set_capacity(herd_size, param_farm)
}
\arguments{
\item{herd_size}{The herd size in a simulated herd.}

\item{param_farm}{See \link{param_farm}.}
}
\value{
Numeric vector of length 2: \code{c(lower_limit_of_herd_size, upper_limit_of_herd_size)}.
}
\description{
Calculate capacity of an area based on inputed parameters
}
