% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_area_table}
\alias{setup_area_table}
\title{Setup of \code{area_table}}
\usage{
setup_area_table(area_table, param)
}
\arguments{
\item{area_table}{See \link{area_table}.}

\item{param}{See \link{param}.}
}
\description{
Setup \link{area_table}.
}
\seealso{
\link{area_table} \link{setup_cows} \link{setup_area_table} \link{setup_movement_table} \link{setup_area_table}
}
