% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_cows}
\alias{setup_cows}
\title{Setup of \code{cow_table}}
\usage{
setup_cows(cow_table, param, save_cows)
}
\arguments{
\item{cow_table}{See \link{cow_table}.}

\item{param}{See \link{param}.}

\item{save_cows}{logical. Whether to save initial \code{cows} to a file.}
}
\value{
A list consisted of \code{init_cows} (\link{cow_table}) as return of the function and \code{month0000.csv} in the directionry specified as \code{param$output_dir}.
}
\description{
Load initial cow status from a csv file, transform it to a \link{cow_table}, and output the \code{cow_table} to a csv file \code{month0000.csv}.
}
\seealso{
\link{cow_table} \link{setup_areas} \link{setup_rp_table} \link{setup_area_table}
}
