% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_rp_table}
\alias{setup_rp_table}
\title{Setup of \code{rp_table}}
\usage{
setup_rp_table(init_n_cows, param_simulation)
}
\arguments{
\item{init_n_cows}{The element \code{init_n_cows} from the return of \code{\link[=setup_cows]{setup_cows()}}.}

\item{param_simulation}{See \link{param_simulation}.}
}
\description{
Make initial \code{rp_table}.
}
\seealso{
\link{setup_cows} \link{setup_areas} \link{rp_table} \link{setup_area_table}
}
