% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{sex_newborns}
\alias{sex_newborns}
\alias{sex_twins}
\title{Sex ratio of newborns}
\usage{
sex_newborns(n_newborns)

sex_twins(n_calves)
}
\arguments{
\item{n_newborns, n_calves}{The number of newborns.}
}
\value{
A character vector consisted of "male", "female" and "freemartin"
}
\description{
\code{sex_newborns()} returns sex of singleton newborns.
\code{sex_twins()} returns sex of pairs of twin newborns.
Probability to be triplets or more is ignored (=0).
}
