% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_blv_spread}
\alias{simulate_blv_spread}
\title{Simulate spread of BLV}
\usage{
simulate_blv_spread(param_simulation, param_farm, param_group,
  list_param_modification = NULL, save_cows = T, save_param = T,
  i_simulation_start = 1)
}
\arguments{
\item{param_simulation}{See \link{param_simulation}.}

\item{param_farm}{See \link{param_farm}.}

\item{param_group}{See \link{param_group}.}

\item{list_param_modification}{List of lists. Parameter specified in each inner list overwrite default parameters. Each inner list is passed to \code{param_modification} of  \code{\link[=calc_param]{calc_param()}}. Specify like \code{list(modification_for_iter1 = list(parameter_name = new_value, ...), modification_for_iter2 = list(...), ...)}.}

\item{save_cows, save_param}{Whether to save \code{result_combined} and \code{param_calculated} (a result of \code{\link[=calc_param]{calc_param()}}) to a file.}

\item{i_simulation_start}{An option to rerun a simulation from the middle of simulations. For example, you run 100 simulation, simulation 26 encounter error and stopped, and you want to run simulation 26-100 again while keeping the result from simulation 1-25. Then set i_simulation = 26.}
}
\value{
The function invisibully returns the result of the final run of simulations. csv files storing cow data and txt files storing parameters information are written to a directory specified by \code{param_simulation$output_dir}.
}
\description{
Simulate spread of BLV in a typical Japanese dairy herd.
}
