% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_once}
\alias{simulate_once}
\title{Body of the simulation}
\usage{
simulate_once(
  cows_areas,
  max_cow_id,
  area_table,
  movement_table,
  day_rp,
  newborn_table,
  i_simulation,
  result,
  param_processed,
  param_modif,
  save_cows,
  save_param
)
}
\arguments{
\item{cows_areas}{A result of \code{\link[=set_init_chamber_id]{set_init_chamber_id()}}.}

\item{max_cow_id}{\code{init_max_cow_id} component of a result of \code{\link[=setup_cows]{setup_cows()}}.}

\item{area_table}{A result of \code{\link[=setup_area_table]{setup_area_table()}}.}

\item{movement_table}{A result of \code{\link[=setup_movement_table]{setup_movement_table()}}.}

\item{day_rp}{A result of \code{\link[=setup_rp_table]{setup_rp_table()}}.}

\item{newborn_table}{A result of \code{\link[=setup_newborn_table]{setup_newborn_table()}}.}

\item{i_simulation}{The iteration number of simulations.}

\item{result}{Lists to store a \code{cow_table}.}

\item{param_processed}{A result of \code{\link[=process_param]{process_param()}}.}

\item{param_modif}{See \code{\link[=calc_param]{calc_param()}}.}

\item{save_cows, save_param}{Wheher to save a result of a simulation and used parameters to files.}
}
\value{
A \code{cow_table}
}
\description{
Run simulation once.
}
\note{
This function does not output the result to a csv file.
}
