% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_function.R
\name{susceptibility}
\alias{susceptibility}
\title{Calculate susceptibility of newborns to the pathogen}
\usage{
susceptibility(
  n_newborns,
  susceptibility_ial_to_ipl_dam,
  susceptibility_ipl_to_ebl_dam,
  param_calculated
)
}
\arguments{
\item{n_newborns}{The number of newborns.}

\item{susceptibility_ial_to_ipl_dam, susceptibility_ipl_to_ebl_dam}{The dam's susceptibility to the pathogen.}

\item{param_calculated}{Return from \code{\link[=calc_param]{calc_param()}}.}
}
\value{
A list consisted of susceptibility of newborns.
}
\description{
Calculate susceptibility of newborns to the pathogen
}
