% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_simulation_process.R
\name{table_route}
\alias{table_route}
\alias{summary_route}
\title{Summarize infection routes}
\usage{
table_route(
  output_filename,
  output_dir,
  n_simulation,
  simulation_length,
  route_levels = NULL
)

summary_route(cows)
}
\arguments{
\item{output_filename}{See \link{param_simulation}.}

\item{output_dir}{See \link{param_simulation}.}

\item{n_simulation}{See \link{param_simulation}.}

\item{simulation_length}{See \link{param_simulation}.}

\item{route_levels}{See \link{redefine_route_levels}.}

\item{cows}{A \code{cow_table} read by \link{read_final_cows}.}
}
\description{
Calculate monthly infection routes at the end of simulations.
}
\seealso{
\link{table_infection_status}
}
