% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barn.R
\name{tether_roaming_cows}
\alias{tether_roaming_cows}
\title{Assign cows roaming in a tie-stall to empty chambers}
\usage{
tether_roaming_cows(cows, areas)
}
\arguments{
\item{cows}{See \link{cow_table}.}

\item{areas}{See \link{setup_area_table}.}
}
\value{
A list consisted of \link{areas} and \link{cow_table}.
}
\description{
Assign \code{chamber_id} to cows roaming in a tie-stall barn.
}
