% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\docType{data}
\name{tie_stall_table}
\alias{tie_stall_table}
\alias{a_chamber}
\title{A data.table to store status of a tie-stall barn}
\format{\link[data.table:data.table]{data.table}}
\usage{
a_chamber
}
\description{
\code{tie_stall_table} is a \link[data.table:data.table]{data.table} to store status of tie-stall barns.
Each tie-stall barn have each \code{tie_stall_table}.
The rows are consisted of \code{a_chamber}, which indicates one chamber in a barn.
}
\details{
\code{chamber_id} and \code{adjoin_previous/next_chamber} are fixed. Values will not be changed while a simulation.
Other variables are flexible. Values will may be changed while a simulation.
\itemize{
\item \code{chamber_id}: Chamber ID.
\item \code{adjoin_previous_chamber}: Whether the chamber adjoins the \code{chamber_id - 1}th chamber.
\item \code{adjoin_next_chamber}: Whether the chamber adjoins the \code{chamber_id + 1}th chamber.
\item \code{cow_id}: Cow ID in a lane.
\item \code{cow_status}: Infection status of the cow.
\item \code{is_exposed}:
TRUE = The cow is not isolated AND one or both of the neighbor cows is infected and not isolated.
FALSE = Both of the neibors are either uninfected or isolated.
NA = No cow in the chamber.
\item \code{is_isolated}: Whether the cow is isolated or not.
\item \code{hazard_ratio}: Hazard ratio of infection to the cow calculated based on neighbors' infection status.
\item \code{previous_neighbor_status}: Infection status of the neighbor in the right chamber.
\item \code{previous_neighbor_isolated}: Whether the cow in the right chamber is isolated or not.
\item \code{previous_neighbor_infectivity}:
TRUE when the neighbor in the right chamber is not isolated and is infectious. Otherwise, FALSE. NA is not allowed to this variable.
\item \code{next_neighbor_status}, \code{next_neighbor_isolated}, \code{next_neighbor_infectivity}: Variables about the neighbor in the left chamber.
}
}
\seealso{
\link{cow_table} \link{area_table} \link{movement_table} \link{rp_table}
}
\keyword{datasets}
